% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_class_BOWL.R
\docType{methods}
\name{BOWL-methods}
\alias{BOWL-methods}
\alias{print,BOWL-method}
\alias{show,BOWL-method}
\title{Methods Available for Objects of Class \code{BOWL}}
\usage{
\S4method{print}{BOWL}(x, ...)

\S4method{show}{BOWL}(object)
}
\description{
Methods Available for Objects of Class \code{BOWL}
}
\keyword{internal}
