% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_class_BOWLBasic.R
\name{BOWLBasic-methods}
\alias{BOWLBasic-methods}
\alias{Call,BOWLBasic-method}
\alias{coef,BOWLBasic-method}
\alias{cvInfo,BOWLBasic-method}
\alias{estimator,BOWLBasic-method}
\alias{fitObject,BOWLBasic-method}
\alias{optimObj,BOWLBasic-method}
\alias{optTx,BOWLBasic,data.frame-method}
\alias{optTx,BOWLBasic,missing-method}
\alias{outcome,BOWLBasic-method}
\alias{plot,BOWLBasic,ANY-method}
\alias{print,BOWLBasic-method}
\alias{propen,BOWLBasic-method}
\alias{regimeCoef,BOWLBasic-method}
\alias{show,BOWLBasic-method}
\alias{summary,BOWLBasic-method}
\title{Methods Available for Objects of Class \code{BOWLBasic}}
\usage{
\S4method{Call}{BOWLBasic}(name, ...)

\S4method{coef}{BOWLBasic}(object, ...)

\S4method{cvInfo}{BOWLBasic}(object, ...)

\S4method{estimator}{BOWLBasic}(x, ...)

\S4method{fitObject}{BOWLBasic}(object, ...)

\S4method{optimObj}{BOWLBasic}(object, ...)

\S4method{optTx}{BOWLBasic,data.frame}(x, newdata, ...)

\S4method{optTx}{BOWLBasic,missing}(x, newdata, ...)

\S4method{outcome}{BOWLBasic}(object, ...)

\S4method{plot}{BOWLBasic,ANY}(x, suppress = FALSE, ...)

\S4method{print}{BOWLBasic}(x, ...)

\S4method{propen}{BOWLBasic}(object, ...)

\S4method{regimeCoef}{BOWLBasic}(object, ...)

\S4method{show}{BOWLBasic}(object)

\S4method{summary}{BOWLBasic}(object, ...)
}
\description{
Methods Available for Objects of Class \code{BOWLBasic}
}
\keyword{internal}
