\name{BOWLObj-class}
\docType{class}
\alias{BOWLObj-class}

\title{Class \code{"BOWLObj"}}

\description{
Class to identify an object as a BOWL step.
}

\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}

\section{Methods}{
No methods defined with class \code{"BOWLObj"} in the signature.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("BOWLObj")
}

\keyword{internal}
