% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_DynTxRegime.R
\name{Call}
\alias{Call}
\title{Retrieve Unevaluated Original Call}
\usage{
Call(name, ...)
}
\arguments{
\item{name}{Object for which call is desired}

\item{...}{Optional additional input required by R's base call().}
}
\description{
Returns the unevaluated original call to a DynTxRegime statistical method.
}
\details{
Methods are defined for all statistical methods implemented in DynTxRegime.
}
