\name{DynTxRegime-class}
\Rdversion{1.0}
\docType{class}
\alias{DynTxRegime-class}

\title{Class \code{"DynTxRegime"}}
\description{A VIRTUAL class containing the results of all DynTxRegime methods.}

\section{Objects from the Class}{
Cannot be created outside of the DynTxRegime package.
}
\author{
  Shannon T. Holloway <sthollow@ncsu.edu>
}

\keyword{classes}
