% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_ClassificationFit.R, R/J_class_OptimalClass.R
\docType{methods}
\name{classif}
\alias{classif}
\alias{classif,OptimalClass-method}
\title{Retrieve Classification Regression Analysis}
\usage{
classif(object, ...)

\S4method{classif}{OptimalClass}(object, ...)
}
\arguments{
\item{object}{Value object returned from a method that uses classification
regression}

\item{...}{Ignored.}
}
\description{
Method retrieves the value object returned by the user specified
  classification regression modeling object(s). Exact structure of the 
  returned object will vary.
}
