% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_ClassificationFit.R
\docType{class}
\name{ClassificationFit-class}
\alias{ClassificationFit-class}
\title{Class \code{ClassificationFit}}
\description{
Class \code{ClassificationFit} combines a \code{TypedFit} object and a 
  \code{TxInfoNoSubsets} object to define a classification regression result
  when subsets are not identified.
}
\keyword{internal}
