% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_ClassificationFit.R
\name{ClassificationFit-methods}
\alias{ClassificationFit-methods}
\alias{classif,ClassificationFit-method}
\alias{coef,ClassificationFit-method}
\alias{fitObject,ClassificationFit-method}
\alias{plot,ClassificationFit,ANY-method}
\alias{predict,ClassificationFit-method}
\alias{.predictAll,ClassificationFit,data.frame-method}
\alias{print,ClassificationFit-method}
\alias{show,ClassificationFit-method}
\alias{summary,ClassificationFit-method}
\title{Methods Available for Objects of Class \code{ClassificationFit}}
\usage{
\S4method{classif}{ClassificationFit}(object, ...)

\S4method{coef}{ClassificationFit}(object, ...)

\S4method{fitObject}{ClassificationFit}(object, ...)

\S4method{plot}{ClassificationFit,ANY}(x, suppress = FALSE, ...)

\S4method{predict}{ClassificationFit}(object, ...)

\S4method{.predictAll}{ClassificationFit,data.frame}(object, newdata, ...)

\S4method{print}{ClassificationFit}(x, ...)

\S4method{show}{ClassificationFit}(object)

\S4method{summary}{ClassificationFit}(object, ...)
}
\description{
\code{.predictAll(object, newdata)}
  predicts optimal treatment
}
\keyword{internal}
