% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_ClassificationFit_fSet.R
\docType{class}
\name{ClassificationFit_fSet-class}
\alias{ClassificationFit_fSet-class}
\title{Class \code{ClassificationFit_fSet}}
\description{
Class \code{ClassificationFit_fSet} contains a \code{TypedFit_fSet} object to
  define a classification regression result when subsets are identified but
  not modeled uniquely.
}
\keyword{internal}
