% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_ClassificationFit_fSet.R
\docType{methods}
\name{ClassificationFit_fSet-methods}
\alias{ClassificationFit_fSet-methods}
\alias{classif,ClassificationFit_fSet-method}
\alias{coef,ClassificationFit_fSet-method}
\alias{fitObject,ClassificationFit_fSet-method}
\alias{.predictAll,ClassificationFit_fSet,data.frame-method}
\alias{summary,ClassificationFit_fSet-method}
\title{Methods Available for Objects of Class \code{ClassificationFit_fSet}}
\usage{
\S4method{classif}{ClassificationFit_fSet}(object, ...)

\S4method{coef}{ClassificationFit_fSet}(object, ...)

\S4method{fitObject}{ClassificationFit_fSet}(object, ...)

\S4method{.predictAll}{ClassificationFit_fSet,data.frame}(object, newdata,
  ...)

\S4method{summary}{ClassificationFit_fSet}(object, ...)
}
\description{
\code{.predictAll(object, newdata)}
  predicts optimal treatment
}
\keyword{internal}
