% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_ClassificationFit_SubsetList.R
\docType{class}
\name{ClassificationFit_SubsetList-class}
\alias{ClassificationFit_SubsetList-class}
\title{Class \code{ClassificationFit_SubsetList}}
\description{
Class \code{ClassificationFit_SubsetList} contains a 
  \code{TypedFit_SubsetList} object to define classification regression
  results when subsets are identified and modeled uniquely.
}
\keyword{internal}
