% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_ClassificationFit_SubsetList.R
\name{ClassificationFit_SubsetList-methods}
\alias{ClassificationFit_SubsetList-methods}
\alias{classif,ClassificationFit_SubsetList-method}
\alias{coef,ClassificationFit_SubsetList-method}
\alias{fitObject,ClassificationFit_SubsetList-method}
\alias{predict,ClassificationFit_SubsetList-method}
\alias{.predictAll,ClassificationFit_SubsetList,data.frame-method}
\alias{summary,ClassificationFit_SubsetList-method}
\title{Methods Available for Objects of Class \code{ClassificationFit_SubsetList}}
\usage{
\S4method{classif}{ClassificationFit_SubsetList}(object, ...)

\S4method{coef}{ClassificationFit_SubsetList}(object, ...)

\S4method{fitObject}{ClassificationFit_SubsetList}(object, ...)

\S4method{predict}{ClassificationFit_SubsetList}(object, ...)

\S4method{.predictAll}{ClassificationFit_SubsetList,data.frame}(object, newdata, ...)

\S4method{summary}{ClassificationFit_SubsetList}(object, ...)
}
\description{
\code{.predictAll(object, newdata)}
  predicts optimal treatment
}
\keyword{internal}
