% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_ClassificationObj.R
\docType{class}
\name{ClassificationObj-class}
\alias{ClassificationObj-class}
\title{Class \code{ClassificationObj}}
\description{
Stores classification regression results under a common name.
}
\section{Slots}{

\describe{
\item{\code{classif}}{ANY - required to be NA, \code{ClassificationFit}, 
\code{ClassificationFit_fSet}, or \cr \code{ClassificationFit_SubsetList}.}
}}

\keyword{internal}
