% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_ClassificationObj.R
\name{ClassificationObj-methods}
\alias{ClassificationObj-methods}
\alias{classif,ClassificationObj-method}
\alias{coef,ClassificationObj-method}
\alias{fitObject,ClassificationObj-method}
\alias{plot,ClassificationObj,ANY-method}
\alias{predict,ClassificationObj-method}
\alias{.predictAll,ClassificationObj,data.frame-method}
\alias{print,ClassificationObj-method}
\alias{show,ClassificationObj-method}
\alias{summary,ClassificationObj-method}
\title{Methods Available for Objects of Class \code{ClassificationObj}}
\usage{
\S4method{classif}{ClassificationObj}(object, ...)

\S4method{coef}{ClassificationObj}(object, ...)

\S4method{fitObject}{ClassificationObj}(object, ...)

\S4method{plot}{ClassificationObj,ANY}(x, suppress = FALSE, ...)

\S4method{predict}{ClassificationObj}(object, ...)

\S4method{.predictAll}{ClassificationObj,data.frame}(object, newdata, ...)

\S4method{print}{ClassificationObj}(x, ...)

\S4method{show}{ClassificationObj}(object)

\S4method{summary}{ClassificationObj}(object, ...)
}
\description{
\code{.predictAll(object, newdata)}
  predicts optimal treatment
}
\keyword{internal}
