% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_generics.R
\name{coef}
\alias{coef}
\title{Extract Model Coefficients From Objects Returned by Modeling Functions}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{Value object returned by any statistical method implemented
in DynTxRegime.}

\item{...}{Optional additional inputs defined by coefficient methods of 
selected regression functions.}
}
\description{
A list is returned, one element for each regression step required by
  the statistical method.
}
\details{
Methods are defined for all statistical methods implemented in DynTxRegime.

The exact structure of the returned list will vary depending on
 the statistical method. 
 For methods that include a propensity regression, the returned list will 
 include an element named 'propen'. For methods that include an outcome 
 regression, the returned list will include an element named 'outcome'.
}
