% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_class_.earl.R
\name{createearl}
\alias{createearl}
\alias{.createearl}
\title{Create method object for EARL}
\usage{
.createearl(kernel, txVec, response, prWgt, surrogate, guess = NULL, mu, ...)
}
\arguments{
\item{kernel}{Kernel object}

\item{txVec}{Vector of tx coded as -1/1}

\item{response}{Vector of responses}

\item{prWgt}{Vector of propensity for tx received}

\item{surrogate}{Surrogate object indicating surrogate loss-function}

\item{guess}{Vector of estimated regime parameters}

\item{mu}{Matrix of outcome regression (zero/ignored)}
}
\value{
An .earl object
}
\description{
Create method object for EARL
}
\keyword{internal}
