% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_class_.owl.R
\name{createowl}
\alias{createowl}
\alias{.createowl}
\title{Create method object for Outcome Weighted Learning}
\usage{
.createowl(kernel, txVec, response, prWgt, surrogate, guess = NULL, mu,
  ...)
}
\arguments{
\item{kernel}{Kernel object}

\item{txVec}{Vector of tx coded as -1/1}

\item{response}{Vector of responses}

\item{prWgt}{Vector of propensity for tx received}

\item{surrogate}{Surrogate object indicating surrogate loss-function}

\item{guess}{Vector of estimated regime parameters}

\item{mu}{Matrix of outcome regression (zero/ignored)}
}
\value{
An .owl object
}
\description{
Create method object for Outcome Weighted Learning
}
\keyword{internal}
