% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_class_.rwl.R
\name{createrwl}
\alias{createrwl}
\alias{.createrwl}
\alias{.createrwlcount}
\title{Create method object for Residual Weighted Learning}
\usage{
.createrwl(kernel, txVec, response, prWgt, surrogate, guess = NULL, mu, ...)

.createrwlcount(
  kernel,
  txVec,
  response,
  prWgt,
  surrogate,
  guess = NULL,
  mu,
  ...
)
}
\arguments{
\item{kernel}{Kernel object}

\item{txVec}{Vector of tx coded as -1/1}

\item{response}{Vector of responses}

\item{prWgt}{Vector of propensity for tx received}

\item{surrogate}{Surrogate object indicating surrogate loss-function}

\item{guess}{Vector of estimated regime parameters}

\item{mu}{Matrix of outcome regression}
}
\value{
An \code{.rwl} object

An \code{.rwl} object
}
\description{
Create method object for Residual Weighted Learning

Create method object for Residual Weighted Learning
}
\keyword{internal}
