% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVBasic.R
\docType{class}
\name{CVBasic-class}
\alias{CVBasic-class}
\title{Class \code{CVBasic}}
\description{
Class \code{CVBasic} holds cross-validation procedure parameters
}
\section{Slots}{

\describe{
\item{\code{folds}}{An integer}

\item{\code{sample}}{A lsit}
}}

\keyword{internal}
