% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVInfo.R
\name{CVInfo-methods}
\alias{CVInfo-methods}
\alias{.getPars,CVInfo-method}
\alias{.getOptimal,CVInfo-method}
\alias{.getValue,CVInfo-method}
\alias{cvInfo,CVInfo-method}
\alias{print,CVInfo-method}
\alias{show,CVInfo-method}
\alias{summary,CVInfo-method}
\title{Methods Available for Objects of Class \code{CVInfo}}
\usage{
\S4method{.getPars}{CVInfo}(object)

\S4method{.getOptimal}{CVInfo}(object)

\S4method{.getValue}{CVInfo}(object)

\S4method{cvInfo}{CVInfo}(object)

\S4method{print}{CVInfo}(x, ...)

\S4method{show}{CVInfo}(object)

\S4method{summary}{CVInfo}(object, ...)
}
\description{
Methods Available for Objects of Class \code{CVInfo}

\code{.getPars}
  retrieves parameters considered in cross-validation.

\code{.getOptimal}
  retrieves optimal parameters identified in cross-validation.

\code{.getValue}
  retrieves values obtained in cross-validation.

\code{cvInfo}
  retrieves cross-validation information.

\code{print}
  print cross-validation results.

\code{show}
  display cross-validation results.

\code{summary}
  summarize cross-validation results.
}
\keyword{internal}
