% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVInfo.R
\name{cvInfo}
\alias{cvInfo}
\title{Extract Cross-Validation Results}
\usage{
cvInfo(object, ...)
}
\arguments{
\item{object}{A value object returned by a weighted learning statistical 
method of DynTxRegime}

\item{...}{Ignored.}
}
\description{
Extract cross-validation results from the value object returned by a 
  weighted learning statistical method of DynTxRegime.
}
\details{
Methods are developed for all weighted learning methods implemented in
 DynTxRegime. Specifically, OWL, RWL, BOWL, and EARL.
}
