% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVInfokParam.R
\docType{class}
\name{CVInfokParam-class}
\alias{CVInfokParam-class}
\title{Class \code{CVInfokParam}}
\description{
Class \code{CVInfokParam} holds information regarding cross-validation
 procedure when only multiple kernel parameters values are considered.
}
\section{Slots}{

\describe{
\item{\code{value}}{Array of values at parameters considered}
}}

\keyword{internal}
