% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVInfoLambda.R
\docType{class}
\name{CVInfoLambda-class}
\alias{CVInfoLambda-class}
\title{Class \code{CVInfoLambda}}
\description{
Class \code{CVInfoLambda} holds information regarding cross-validation
  procedure when only multiple lambda values are considered.
}
\section{Slots}{

\describe{
\item{\code{value}}{Array of values at tuning parameters considered}
}}

\keyword{internal}
