% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVInfoObj.R
\docType{class}
\name{CVInfoObj-class}
\alias{CVInfoObj-class}
\title{Class \code{CVInfoObj}}
\description{
Class \code{CVInfoObj} holds information regarding cross-validation
  procedure under a common name.
}
\section{Slots}{

\describe{
\item{\code{cvInfo}}{ANY expected to be CVInfo or NULL}
}}

\keyword{internal}
