% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVInfoObj.R
\name{CVInfoObj-methods}
\alias{CVInfoObj-methods}
\alias{.getPars,CVInfoObj-method}
\alias{.getOptimal,CVInfoObj-method}
\alias{.getValue,CVInfoObj-method}
\alias{cvInfo,CVInfoObj-method}
\alias{print,CVInfoObj-method}
\alias{show,CVInfoObj-method}
\alias{summary,CVInfoObj-method}
\title{Methods Available for Objects of Class \code{CVInfoObj}}
\usage{
\S4method{.getPars}{CVInfoObj}(object)

\S4method{.getOptimal}{CVInfoObj}(object)

\S4method{.getValue}{CVInfoObj}(object)

\S4method{cvInfo}{CVInfoObj}(object)

\S4method{print}{CVInfoObj}(x, ...)

\S4method{show}{CVInfoObj}(object)

\S4method{summary}{CVInfoObj}(object, ...)
}
\description{
Call methods equivalently named for object inheriting from CVInfo.
Methods dispached depend on object in @cvInfo.
}
\keyword{internal}
