% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_DecisionPointList.R
\name{DecisionPointList-methods}
\alias{DecisionPointList-methods}
\alias{plot,DecisionPointList,ANY-method}
\alias{print,DecisionPointList-method}
\alias{show,DecisionPointList-method}
\title{Methods Available for Objects of Class \code{DecisionPointList}}
\usage{
\S4method{plot}{DecisionPointList,ANY}(x, suppress = FALSE, ...)

\S4method{print}{DecisionPointList}(x, ...)

\S4method{show}{DecisionPointList}(object)
}
\description{
Methods Available for Objects of Class \code{DecisionPointList}

\code{plot(x,suppress)}
  generates plots of the regression analysis for each decision point.
  If suppress = FALSE, titles of plot will include the decision point
  identifier.

\code{print(x)}
  adds decision point information to print statements.
  Each decision point is preceded by 'dp=x' where x is the decision point
  number.

\code{show(object)}
  adds decision point information to show statements.
  Each decision point is preceded by 'dp=x' where x is the decision point
  number.
}
\keyword{internal}
