% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_class_OptimalClass.R
\name{.optimalClass}
\alias{.optimalClass}
\title{Perform Classification Step}
\usage{
.optimalClass(moPropen, moMain, moCont, moClass, data, response, txName,
  iter, fSet, suppress, step)
}
\arguments{
\item{moPropen}{model object(s) for propensity regression}

\item{moMain}{model object(s) for main effects of outcome regression or NULL}

\item{moCont}{model object(s) for contrasts of outcome regression or NULL}

\item{moClass}{model object(s) for classification procedure}

\item{data}{data.frame of covariates and treatment history}

\item{response}{vector of responses}

\item{txName}{character of column header of data containing tx}

\item{iter}{maximum number of iterations for outcome regression or NULL}

\item{fSet}{function defining subsets or NULL}

\item{suppress}{T/F indicating screen printing preference}

\item{step}{integer indicating step of algorithm}
}
\value{
an object of class OptimalClass
}
\description{
Perform Classification Step
}
\keyword{internal}
