% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_QLearn.R
\name{DTRstep}
\alias{DTRstep}
\title{Identify Statistical Method Used to Obtain Result}
\usage{
DTRstep(object)
}
\arguments{
\item{object}{Value object returned by any statistical method of DynTxRegime}
}
\description{
Prints are displays a brief description of the statistical method used to
  obtain the input object.
}
\details{
Methods are defined for all statistical methods implemented in DynTxRegime.
}
