\name{internals}




\alias{DTRstep,DynTxRegime-method}
\alias{classif,DynTxRegime-method}
\alias{cvInfo,DynTxRegime-method}
\alias{fittedCont,DynTxRegime-method}
\alias{fittedMain,DynTxRegime-method}
\alias{genetic,DynTxRegime-method}
\alias{optimObj,DynTxRegime-method}
\alias{regimeCoef,DynTxRegime-method}
\alias{residuals,DynTxRegime-method}
\alias{sd,DynTxRegime-method}

\alias{DTRstep,BOWL-method}
\alias{cvInfo,BOWL-method}
\alias{optimObj,BOWL-method}
\alias{optTx,BOWL,missing-method}
\alias{optTx,BOWL,data.frame-method}
\alias{print,BOWL-method}
\alias{regimeCoef,BOWL-method}
\alias{show,BOWL-method}
\alias{summary,BOWL-method}

\alias{DTRstep,EARLIPWE-method}
\alias{print,EARLIPWE-method}
\alias{show,EARLIPWE-method}
\alias{summary,EARLIPWE-method}

\alias{DTRstep,EARLAIPWE-method}
\alias{print,EARLAIPWE-method}
\alias{show,EARLAIPWE-method}
\alias{summary,EARLAIPWE-method}


\alias{fitObject,OptimalClassAIPWE-method}
\alias{fitObject,OptimalClassIPWE-method}

\alias{DTRstep,IQLearnFS_C-method}
\alias{print,IQLearnFS_C-method}
\alias{show,IQLearnFS_C-method}
\alias{residuals,IQLearnFS_C-method}

\alias{DTRstep,IQLearnFS_ME-method}
\alias{print,IQLearnFS_ME-method}
\alias{show,IQLearnFS_ME-method}

\alias{DTRstep,IQLearnFS_VHet-method}
\alias{print,IQLearnFS_VHet-method}
\alias{show,IQLearnFS_VHet-method}
\alias{residuals,IQLearnFS_VHet-method}

\alias{DTRstep,IQLearnSS-method}
\alias{estimator,IQLearnSS-method}
\alias{fittedCont,IQLearnSS-method}
\alias{fittedMain,IQLearnSS-method}
\alias{print,IQLearnSS-method}
\alias{optTx,IQLearnSS,missing-method}
\alias{optTx,IQLearnSS,data.frame-method}
\alias{show,IQLearnSS-method}

\alias{classif,OptimalClassIPWE-method}
\alias{DTRstep,OptimalClassIPWE-method}
\alias{fitObject,OptimalClassIPWE-method}
\alias{optTx,OptimalClassIPWE,data.frame-method}
\alias{optTx,OptimalClassIPWE,missing-method}
\alias{plot,OptimalClassIPWE,ANY-method}
\alias{print,OptimalClassIPWE-method}
\alias{show,OptimalClassIPWE-method}
\alias{summary,OptimalClassIPWE-method}

\alias{classif,OptimalClassAIPWE-method}
\alias{DTRstep,OptimalClassAIPWE-method}
\alias{fitObject,OptimalClassAIPWE-method}
\alias{optTx,OptimalClassAIPWE,data.frame-method}
\alias{optTx,OptimalClassAIPWE,missing-method}
\alias{plot,OptimalClassAIPWE,ANY-method}
\alias{print,OptimalClassAIPWE-method}
\alias{show,OptimalClassAIPWE-method}
\alias{summary,OptimalClassAIPWE-method}

\alias{DTRstep,OptimalSeqIPWE_SDP-method}
\alias{genetic,OptimalSeqIPWE_SDP-method}
\alias{optTx,OptimalSeqIPWE_SDP,data.frame-method}
\alias{print,OptimalSeqIPWE_SDP-method}
\alias{regimeCoef,OptimalSeqIPWE_SDP-method}
\alias{show,OptimalSeqIPWE_SDP-method}
\alias{summary,OptimalSeqIPWE_SDP-method}

\alias{DTRstep,OptimalSeqAIPWE_SDP-method}
\alias{genetic,OptimalSeqAIPWE_SDP-method}
\alias{optTx,OptimalSeqAIPWE_SDP,data.frame-method}
\alias{print,OptimalSeqAIPWE_SDP-method}
\alias{regimeCoef,OptimalSeqAIPWE_SDP-method}
\alias{show,OptimalSeqAIPWE_SDP-method}
\alias{summary,OptimalSeqAIPWE_SDP-method}

\alias{DTRstep,OptimalSeqIPWE_MDP-method}
\alias{genetic,OptimalSeqIPWE_MDP-method}
\alias{optTx,OptimalSeqIPWE_MDP,data.frame-method}
\alias{optTx,OptimalSeqIPWE_MDP,missing-method}
\alias{print,OptimalSeqIPWE_MDP-method}
\alias{regimeCoef,OptimalSeqIPWE_MDP-method}
\alias{show,OptimalSeqIPWE_MDP-method}
\alias{summary,OptimalSeqIPWE_MDP-method}

\alias{DTRstep,OptimalSeqAIPWE_MDP-method}
\alias{genetic,OptimalSeqAIPWE_MDP-method}
\alias{optTx,OptimalSeqAIPWE_MDP,data.frame-method}
\alias{optTx,OptimalSeqAIPWE_MDP,missing-method}
\alias{print,OptimalSeqAIPWE_MDP-method}
\alias{regimeCoef,OptimalSeqAIPWE_MDP-method}
\alias{show,OptimalSeqAIPWE_MDP-method}
\alias{summary,OptimalSeqAIPWE_MDP-method}

\alias{DTRstep,OWL-method}
\alias{cvInfo,OWL-method}
\alias{optimObj,OWL-method}
\alias{optTx,OWL,data.frame-method}
\alias{optTx,OWL,missing-method}
\alias{print,OWL-method}
\alias{regimeCoef,OWL-method}
\alias{show,OWL-method}
\alias{summary,OWL-method}

\alias{DTRstep,QLearn-method}
\alias{optTx,QLearn,data.frame-method}
\alias{optTx,QLearn,missing-method}
\alias{print,QLearn-method}
\alias{show,QLearn-method}
\alias{summary,QLearn-method}

\alias{cvInfo,EARL-method}
\alias{optimObj,EARL-method}
\alias{optTx,EARL,missing-method}
\alias{optTx,EARL,data.frame-method}
\alias{print,EARL-method}
\alias{regimeCoef,EARL-method}
\alias{show,EARL-method}
\alias{summary,EARL-method}

\alias{regimeCoef,EARLOptim-method}
\alias{regimeCoef,OWLOptim-method}
\alias{regimeCoef,RWLOptim-method}

\alias{plot,PropensityFit,ANY-method}

\alias{coef,PropensityAndOutcome-method}
\alias{fitObject,PropensityAndOutcome-method}
\alias{outcome,PropensityAndOutcome-method}
\alias{plot,PropensityAndOutcome,ANY-method}
\alias{print,PropensityAndOutcome-method}
\alias{propen,PropensityAndOutcome-method}
\alias{show,PropensityAndOutcome-method}
\alias{summary,PropensityAndOutcome-method}

\alias{coef,PropensityOnly-method}
\alias{fitObject,PropensityOnly-method}
\alias{outcome,PropensityOnly-method}
\alias{plot,PropensityOnly,ANY-method}
\alias{print,PropensityOnly-method}
\alias{propen,PropensityOnly-method}
\alias{show,PropensityOnly-method}
\alias{summary,PropensityOnly-method}

\alias{coef,OutcomeOnly-method}
\alias{fitObject,OutcomeOnly-method}
\alias{outcome,OutcomeOnly-method}
\alias{plot,OutcomeOnly,ANY-method}
\alias{print,OutcomeOnly-method}
\alias{propen,OutcomeOnly-method}
\alias{show,OutcomeOnly-method}
\alias{summary,OutcomeOnly-method}

\alias{DTRstep,RWL-method}
\alias{cvInfo,RWL-method}
\alias{optTx,RWL,missing-method}
\alias{optTx,RWL,data.frame-method}
\alias{optimObj,RWL-method}
\alias{print,RWL-method}
\alias{regimeCoef,RWL-method}
\alias{residuals,RWL-method}
\alias{show,RWL-method}
\alias{summary,RWL-method}

\alias{sd,ANY-method}

\title{ ~~ }
\description{~~}
\keyword{internal}
