% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_DynTxRegime.R
\name{DynTxRegime-methods}
\alias{DynTxRegime-methods}
\alias{Call,DynTxRegime-method}
\title{Methods Available for Objects of Class \code{DynTxRegime}}
\usage{
\S4method{Call}{DynTxRegime}(name, ...)
}
\description{
Methods Available for Objects of Class \code{DynTxRegime}

\code{Call(name)}
  retrieves the unevaluated call to the original statistical method
}
\keyword{internal}
