% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_OptimalInfo.R, R/E_class_IQLearnFS.R
\name{estimator}
\alias{estimator}
\alias{estimator,IQLearnFS-method}
\alias{estimator,IQLearnSS-method}
\title{Retrieve the Estimated Value}
\usage{
estimator(x, ...)

\S4method{estimator}{IQLearnFS}(x, w = NULL, y = NULL, z = NULL, dens = NULL)

\S4method{estimator}{IQLearnSS}(x, w = NULL, y = NULL, z = NULL, dens = NULL)
}
\arguments{
\item{x}{a DynTxRegime Object.}

\item{...}{Optional additional input. Ignored.}

\item{w}{If IQ-Learning, object of class IQLearnSS, IQLearnFS_C, 
IQLearnFS_ME, or IQLearnFS_VHet}

\item{y}{If IQ-Learning, object of class IQLearnSS, IQLearnFS_C, 
IQLearnFS_ME, or IQLearnFS_VHet}

\item{z}{If IQ-Learning, object of class IQLearnSS, IQLearnFS_C, 
IQLearnFS_ME, or IQLearnFS_VHet}

\item{dens}{If IQ-Learning, one of {norm, nonpar}}
}
\description{
Retrieve the value as estimated by the statistical method.
}
