% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_ExpSurrogate.R
\docType{methods}
\name{ExpSurrogate-methods}
\alias{ExpSurrogate-methods}
\alias{.phiFunc,ExpSurrogate-method}
\alias{.dPhiFunc,ExpSurrogate-method}
\title{Methods Available for Objects of Class \code{ExpSurrogate}}
\usage{
\S4method{.phiFunc}{ExpSurrogate}(surrogate, u)

\S4method{.dPhiFunc}{ExpSurrogate}(surrogate, u, du)
}
\description{
Methods Available for Objects of Class \code{ExpSurrogate}

\code{.phiFunc}
  calculates exponential surrogate loss-function

\code{.dphiFunc}
  calculates derivative of exponential surrogate loss-function
}
\keyword{internal}
