% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_generics.R
\name{fitObject}
\alias{fitObject}
\title{Objects Returned by Modeling Functions}
\usage{
fitObject(object, ...)
}
\arguments{
\item{object}{Value object returned by a statistical method of DynTxRegime}

\item{...}{Optional additional inputs}
}
\description{
Returns a list of the objects returned by all modeling functions
}
\details{
Methods are defined for all statistical methods implemented in DynTxRegime.

The exact structure of the returned list will vary depending on
 the statistical method. 
 For methods that include a propensity regression, the returned list will 
 include an element named 'propen'. For methods that include an outcome 
 regression, the returned list will include an element named 'outcome'.
}
