% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnSS.R
\docType{methods}
\name{fittedCont}
\alias{fittedCont}
\alias{fittedCont,IQLearnSS-method}
\title{Retrieve the Fitted Contrast Component from Second Stage IQ-Learning}
\usage{
fittedCont(object, ...)

\S4method{fittedCont}{IQLearnSS}(object, ...)
}
\arguments{
\item{object}{An object of class IQLearnSS}

\item{...}{Ignored.}
}
\description{
Extracts the contrasts component of the fitted outcome regression
 the second-stage analysis of the interactive Q-Learning algorithm.
}
