% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnSS.R
\docType{methods}
\name{fittedMain}
\alias{fittedMain}
\alias{fittedMain,IQLearnSS-method}
\title{Retrieve the Fitted Main Effects Component from Second Stage IQ-Learning}
\usage{
fittedMain(object, ...)

\S4method{fittedMain}{IQLearnSS}(object, ...)
}
\arguments{
\item{object}{An object of class IQLearnSS}

\item{...}{Ignored.}
}
\description{
Extracts the main effects component of the fitted outcome regression for
 the second-stage analysis of the interactive Q-Learning algorithm.
}
