% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_class_OptimalSeq.R
\docType{methods}
\name{genetic}
\alias{genetic}
\alias{genetic,OptimalSeq-method}
\title{Retrieve the Genetic Algorithm Results}
\usage{
genetic(object, ...)

\S4method{genetic}{OptimalSeq}(object, ...)
}
\arguments{
\item{object}{Value object returned by optimalSeq()}

\item{...}{Optional inputs. Ignored.}
}
\description{
Retrieve the value object returned by rgenoud() in optimalSeq().
}
