% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_class_OptimalClass.R
\name{getOutcome}
\alias{getOutcome}
\alias{.getOutcome}
\alias{.getOutcome2}
\title{Retrieve Outcome for Both Tx Options When Tx is Binary}
\usage{
.getOutcome(outcomeObj, txObj, data)

.getOutcome2(outcomeObj, txObj, data)
}
\arguments{
\item{outcomeObj}{a OutcomeObj}

\item{txObj}{a TxObj}

\item{data}{a data.frame}
}
\value{
matrix of outcome under binary tx.

matrix of outcome under binary tx.
}
\description{
Retrieve Outcome for Both Tx Options When Tx is Binary

Retrieve Outcome for Both Tx Options When Tx is Binary
}
\keyword{internal}
