% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_class_OptimalSeqMissing.R
\name{getPrWgt}
\alias{getPrWgt}
\alias{.getPrWgt}
\title{Retrieve Propensity for Tx Received}
\usage{
.getPrWgt(propenObj, txObj, data)
}
\arguments{
\item{propenObj}{a PropensityObj}

\item{txObj}{a TxObj}

\item{data}{a data.frame}
}
\value{
vector of propensity for tx received.
}
\description{
Retrieve Propensity for Tx Received
}
\keyword{internal}
