% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_HingeSurrogate.R
\name{HingeSurrogate-methods}
\alias{HingeSurrogate-methods}
\alias{.phiFunc,HingeSurrogate-method}
\alias{.dPhiFunc,HingeSurrogate-method}
\alias{.optim,HingeSurrogate-method}
\title{Methods Available for Objects of Class \code{HingeSurrogate}}
\usage{
\S4method{.phiFunc}{HingeSurrogate}(surrogate, u)

\S4method{.dPhiFunc}{HingeSurrogate}(surrogate, u, du)

\S4method{.optim}{HingeSurrogate}(surrogate, par, lambda, fn, gr, suppress, ...)
}
\description{
Utilizes dfoptim::hjk to obtain parameter estimates. Requires that
the objective function be defined by the calling learning method.
Returns NULL if optimization is not successful due to problems or
the list object returned by dfoptim::hjk if optimization is successful.
}
\keyword{internal}
