% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_HuberHingeSurrogate.R
\docType{class}
\name{HuberHingeSurrogate-class}
\alias{HuberHingeSurrogate-class}
\title{Class \code{HuberHingeSurrogate}}
\description{
Huberized hinge surrogate for 0/1 loss function.
}
\keyword{internal}
