% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_HuberHingeSurrogate.R
\name{HuberHingeSurrogate-methods}
\alias{HuberHingeSurrogate-methods}
\alias{.phiFunc,HuberHingeSurrogate-method}
\alias{.dPhiFunc,HuberHingeSurrogate-method}
\title{Methods Available for Objects of Class \code{HuberHingeSurrogate}}
\usage{
\S4method{.phiFunc}{HuberHingeSurrogate}(surrogate, u)

\S4method{.dPhiFunc}{HuberHingeSurrogate}(surrogate, u, du)
}
\description{
Methods Available for Objects of Class \code{HuberHingeSurrogate}

\code{.phiFunc}
  calculates huberized hinge surrogate loss-function

\code{.dphiFunc}
  calculates derivative of huberized hinge surrogate loss-function
}
\keyword{internal}
