% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_class_.earl.R
\docType{class}
\name{internal-earl-class}
\alias{internal-earl-class}
\title{Class \code{.earl}}
\description{
Class \code{.earl} stores parameters required for EARL optimization step.
}
\section{Slots}{

\describe{
\item{\code{x}}{Matrix of covariates for kernel}

\item{\code{wp}}{Vector of positive weights}

\item{\code{wn}}{Vector of negative weights}

\item{\code{mu}}{Matrix of outcome regression}

\item{\code{txVec}}{Vector of treatment coded as -1/1}

\item{\code{invPi}}{Vector of inverse propensity for treatment received}

\item{\code{response}}{Vector of the response}

\item{\code{surrogate}}{The Surrogate for the loss-function}

\item{\code{par}}{Vector of regime parameters}

\item{\code{kernel}}{The Kernel defining the decision function}
}}

\keyword{internal}
