% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_class_.earl.R
\name{internal-earl-methods}
\alias{internal-earl-methods}
\alias{.subsetObject,.earl-method}
\alias{.objFn,numeric,.earl,LinearKernel-method}
\alias{.dobjFn,numeric,.earl,LinearKernel-method}
\alias{.objFn,numeric,.earl,Kernel-method}
\alias{.dobjFn,numeric,.earl,Kernel-method}
\alias{.objFn,numeric,.earl,MultiRadialKernel-method}
\alias{.dobjFn,numeric,.earl,MultiRadialKernel-method}
\alias{.valueFunc,.earl-method}
\title{Methods Available for Objects of Class \code{.earl}}
\usage{
\S4method{.subsetObject}{.earl}(methodObject, subset)

\S4method{.objFn}{numeric,.earl,LinearKernel}(par, methodObject, kernel, lambda, ...)

\S4method{.dobjFn}{numeric,.earl,LinearKernel}(par, methodObject, kernel, lambda, ...)

\S4method{.objFn}{numeric,.earl,Kernel}(par, methodObject, kernel, lambda, ...)

\S4method{.dobjFn}{numeric,.earl,Kernel}(par, methodObject, kernel, lambda, ...)

\S4method{.objFn}{numeric,.earl,MultiRadialKernel}(par, methodObject, kernel, lambda, ...)

\S4method{.dobjFn}{numeric,.earl,MultiRadialKernel}(par, methodObject, kernel, lambda, ...)

\S4method{.valueFunc}{.earl}(methodObject, optTx, ...)
}
\description{
Methods Available for Objects of Class \code{.earl}

\code{.objFn} not allowed for EARL with multiple radial kernels

\code{.dobjFn} not allowed for EARL with multiple radial kernels
}
\keyword{internal}
