% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_class_.owl.R
\docType{methods}
\name{internal-owl-methods}
\alias{internal-owl-methods}
\alias{.subsetObject,.owl-method}
\alias{.objFn,numeric,.owl,Kernel-method}
\alias{.dobjFn,numeric,.owl,Kernel-method}
\alias{.objFn,numeric,.owl,LinearKernel-method}
\alias{.dobjFn,numeric,.owl,LinearKernel-method}
\alias{.objFn,numeric,.owl,MultiRadialKernel-method}
\alias{.dobjFn,numeric,.owl,MultiRadialKernel-method}
\alias{.valueFunc,.owl-method}
\title{Methods Available for Objects of Class \code{.owl}}
\usage{
\S4method{.subsetObject}{.owl}(methodObject, subset)

\S4method{.objFn}{numeric,.owl,Kernel}(par, methodObject, kernel, lambda,
  ...)

\S4method{.dobjFn}{numeric,.owl,Kernel}(par, methodObject, kernel, lambda,
  ...)

\S4method{.objFn}{numeric,.owl,LinearKernel}(par, methodObject, kernel,
  lambda, ...)

\S4method{.dobjFn}{numeric,.owl,LinearKernel}(par, methodObject, kernel,
  lambda, ...)

\S4method{.objFn}{numeric,.owl,MultiRadialKernel}(par, methodObject,
  kernel, lambda, ...)

\S4method{.dobjFn}{numeric,.owl,MultiRadialKernel}(par, methodObject,
  kernel, lambda, ...)

\S4method{.valueFunc}{.owl}(methodObject, optTx, ...)
}
\description{
Methods Available for Objects of Class \code{.owl}

code{.objFn} is not allowed for OWL with multiple radial kernels

code{.dobjFn} is not allowed for OWL with multiple radial kernels
}
\keyword{internal}
