% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_class_.rwl.R
\docType{class}
\name{internal-rwl-class}
\alias{internal-rwl-class}
\title{Class \code{.rwl}}
\description{
Class \code{.rwl} stores parameters required for an RWL optimization step
}
\section{Slots}{

\describe{
\item{\code{x}}{Matrix of covariates for kernel}

\item{\code{txVec}}{Vector of treatment coded as -1/1}

\item{\code{absRinvPi}}{Vector of the absolute value of the residual weighted by
the propensity for the treatment received}

\item{\code{residual}}{Vector of the residuals}

\item{\code{response}}{Vector of the response}

\item{\code{beta}}{Vector of beta parameters}

\item{\code{surrogate}}{The Surrogate for the loss-function}

\item{\code{pars}}{Vector of regime parameters}

\item{\code{kernel}}{The Kernel defining the decision function}
}}

\keyword{internal}
