% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_class_.rwl.R
\docType{methods}
\name{internal-rwl-methods}
\alias{internal-rwl-methods}
\alias{.subsetObject,.rwl-method}
\alias{.objFn,numeric,.rwl,LinearKernel-method}
\alias{.dobjFn,numeric,.rwl,LinearKernel-method}
\alias{.objFn,numeric,.rwl,Kernel-method}
\alias{.dobjFn,numeric,.rwl,Kernel-method}
\alias{.objFn,numeric,.rwl,MultiRadialKernel-method}
\alias{.dobjFn,numeric,.rwl,MultiRadialKernel-method}
\alias{.valueFunc,.rwl-method}
\title{Methods Available for Objects of Class \code{.rwl}}
\usage{
\S4method{.subsetObject}{.rwl}(methodObject, subset)

\S4method{.objFn}{numeric,.rwl,LinearKernel}(par, methodObject, kernel,
  lambda, ...)

\S4method{.dobjFn}{numeric,.rwl,LinearKernel}(par, methodObject, kernel,
  lambda, ...)

\S4method{.objFn}{numeric,.rwl,Kernel}(par, methodObject, kernel, lambda,
  ...)

\S4method{.dobjFn}{numeric,.rwl,Kernel}(par, methodObject, kernel, lambda,
  ...)

\S4method{.objFn}{numeric,.rwl,MultiRadialKernel}(par, methodObject,
  kernel, lambda, ...)

\S4method{.dobjFn}{numeric,.rwl,MultiRadialKernel}(par, methodObject,
  kernel, lambda, ...)

\S4method{.valueFunc}{.rwl}(methodObject, optTx, ...)
}
\description{
Methods Available for Objects of Class \code{.rwl}

\code{.objFn}
  not allowed for RWL With multiple radial kernels

\code{.dobjFn}
  not allowed for RWL With multiple radial kernels
}
\keyword{internal}
