\name{IQLearnBase-class}
\docType{class}
\alias{IQLearnBase-class}

\title{Class \code{"IQLearnBase"}}
\description{
Regression result for an IQ-Learning step.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("IQLearnBase", ...)}.
These objects are for convenience in package development
and should not be created by users.
}

\section{Extends}{
Class \code{"\link{OutcomeOnly-class},"} directly.
}
\section{Methods}{
  No methods with signature \code{"IQLearnBase"} identified.
}
\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("IQLearnBase")
}

\keyword{internal}


