% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnFS.R
\docType{class}
\name{IQLearnFS-class}
\alias{IQLearnFS-class}
\title{Class \code{IQLearnFS}}
\description{
Class \code{IQLearnFS} contains results for a component of the first stage
  analysis of the interactive Q-Learning algorithm.
}
\keyword{internal}
