% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnFS.R
\docType{methods}
\name{IQLearnFS-methods}
\alias{IQLearnFS-methods}
\alias{print,IQLearnFS-method}
\alias{show,IQLearnFS-method}
\alias{summary,IQLearnFS-method}
\title{Methods Available for Objects of Class \code{IQLearnFS}}
\usage{
\S4method{print}{IQLearnFS}(x, ...)

\S4method{show}{IQLearnFS}(object)

\S4method{summary}{IQLearnFS}(object, ...)
}
\description{
Employs methods defined for \code{QLearn}
}
\keyword{internal}
