% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnFS_VHet.R
\docType{methods}
\name{IQLearnFS_VHet-methods}
\alias{IQLearnFS_VHet-methods}
\alias{print,IQLearnFS_VHet-method}
\alias{show,IQLearnFS_VHet-method}
\title{Methods Available for Objects of Class \code{IQLearnFS_VHet}}
\usage{
\S4method{print}{IQLearnFS_VHet}(x, ...)

\S4method{show}{IQLearnFS_VHet}(object)
}
\description{
Methods Available for Objects of Class \code{IQLearnFS_VHet}
}
\keyword{internal}
