% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnSS.R
\docType{class}
\name{IQLearnSS-class}
\alias{IQLearnSS-class}
\title{Class \code{IQLearnSS}}
\description{
Class \code{IQLearnSS} contains the results for the second stage
  of the interactive Q-Learning algorithm. Objects of this class are
  returned by iqLearnSS().
}
\section{Slots}{

\describe{
\item{\code{yContHat}}{: A numeric. Estimated contrast component}

\item{\code{yMainHat}}{: A numeric. Estimated main effects component}

\item{\code{delta}}{: A numeric. Indicator of compliance * response used for value calc}

\item{\code{step}}{: Not used in this context.}

\item{\code{outcome}}{: The outcome regression analysis}

\item{\code{txInfo}}{: The feasible tx information}

\item{\code{optimal}}{: The estimated optimal tx, decision function, and value}
}}

\section{Methods For Post-Processing of Regression Analysis}{


 \describe{
   \item{outcome}{: 
      Retrieve value object returned by outcome regression methods.  }
}



 \describe{
   \item{coef}{: 
     Retrieve parameter estimates for all regression steps. }
   \item{fitObject}{: 
      Retrieve value object returned by regression methods. }
   \item{plot}{: 
      Generate plots for regression analyses.  }
}


\describe{
   \item{fittCont}{:Retrieve the contrasts component of the regression.}
   \item{fittMain}{:Retrieve the main effects component of the regression.}
}
}

\section{Methods For Accessing Main Results}{


 \describe{
   \item{DTRstep}{: 
      Retrieve description of method used to create object. }
   \item{estimator}{: 
      Retrieve the estimated value of the estimated
      optimal regime for the training data set. }
   \item{optTx}{: 
      Retrieve/predict the estimated decision functions and/or optimal tx. }
   \item{print}{: 
      Print main results of analysis.}
   \item{show}{: 
      Show main results of analysis.}
   \item{summary}{: 
      Retrieve summary information.}
}
}

