\name{IterateFitNoSubsets-class}
\docType{class}
\alias{IterateFitNoSubsets-class}

\title{Class \code{"IterateFitNoSubsets"}}

\description{
An outcome regression step with a single model and no
subset modeling.
}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("IterateFitNoSubsets", ...)}.
These objects are for convenience in package development
and should not be created by users.
}

\section{Extends}{
Class \code{"\link{IterateFit-class}"} directly.
Classes \code{"\link{SingleDecisionPoint-class}"}, 
\code{"\link{OutcomeRegression-class}"}, and \code{"\link{SubsetsNotModeled-class}"}
by class \code{"\link{IterateFit-class}"} distance 2.
}

\section{Methods}{
No methods defined with class \code{"IterateFitNoSubsets"} in the signature.
}

\author{
Shannon T Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("IterateFitNoSubsets")
}

\keyword{internal}
