% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_Kernel.R
\docType{class}
\name{Kernel-class}
\alias{Kernel-class}
\title{Class \code{Kernel}}
\description{
Class \code{Kernel} holds information regarding the decision function kernel
}
\section{Slots}{

\describe{
\item{\code{model}}{An formula. Defines the covariates of the kernel.}

\item{\code{X}}{A matrix. The covariates of the kernel}

\item{\code{kparam}}{ANY. The kernel parameter}
}}

\keyword{internal}
