% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_Kernel.R
\docType{methods}
\name{Kernel-methods}
\alias{Kernel-methods}
\alias{.getKernelX,data.frame,Kernel-method}
\alias{.kernelNumPars,Kernel-method}
\alias{.kernel,Kernel,missing,missing-method}
\alias{.kernel,Kernel,data.frame,data.frame-method}
\alias{.kernel,Kernel,vector,vector-method}
\alias{.kernel,Kernel,vector,data.frame-method}
\alias{.kernel,Kernel,data.frame,vector-method}
\alias{.kernel,Kernel,matrix,data.frame-method}
\alias{.kernel,Kernel,data.frame,matrix-method}
\alias{.kernel,Kernel,vector,matrix-method}
\alias{.kernel,Kernel,matrix,vector-method}
\alias{.kernel,Kernel,matrix,matrix-method}
\alias{print,Kernel-method}
\alias{show,Kernel-method}
\alias{summary,Kernel-method}
\title{Methods Available for Objects of Class \code{Kernel}}
\usage{
\S4method{.getKernelX}{data.frame,Kernel}(data, object)

\S4method{.kernelNumPars}{Kernel}(object, ...)

\S4method{.kernel}{Kernel,missing,missing}(object, x1, x2, ...)

\S4method{.kernel}{Kernel,data.frame,data.frame}(object, x1, x2, ...)

\S4method{.kernel}{Kernel,vector,vector}(object, x1, x2, ...)

\S4method{.kernel}{Kernel,vector,data.frame}(object, x1, x2, ...)

\S4method{.kernel}{Kernel,data.frame,vector}(object, x1, x2, ...)

\S4method{.kernel}{Kernel,matrix,data.frame}(object, x1, x2, ...)

\S4method{.kernel}{Kernel,data.frame,matrix}(object, x1, x2, ...)

\S4method{.kernel}{Kernel,vector,matrix}(object, x1, x2, ...)

\S4method{.kernel}{Kernel,matrix,vector}(object, x1, x2, ...)

\S4method{.kernel}{Kernel,matrix,matrix}(object, x1, x2, ...)

\S4method{print}{Kernel}(x, ...)

\S4method{show}{Kernel}(object)

\S4method{summary}{Kernel}(object, ...)
}
\description{
Methods Available for Objects of Class \code{Kernel}

\code{.getKernelX}
  retrieves the covariates matrix of the kernel.

\code{.kernelNumPars}
  retrieves the number of covariates of the kernel.

\code{.kernel}
  calculates the kernel

\code{print}
  prints kernel model.

\code{show}
  displays kernel model.

\code{summary}
  returns a list containing the kernel model.
}
\keyword{internal}
