% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_KernelObj.R
\docType{class}
\name{KernelObj-class}
\alias{KernelObj-class}
\alias{.kernelNumPars,KernelObj-method}
\title{Class \code{KernelObj}}
\usage{
\S4method{.kernelNumPars}{KernelObj}(object, ...)
}
\description{
Class \code{KernelObj} holds decision function kernel information
  under a common name.
}
\section{Slots}{

\describe{
\item{\code{kernel}}{ANY expected to be \code{Kernel} or \code{SubsetList}}
}}

\keyword{internal}
