% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_KernelObj.R
\docType{methods}
\name{KernelObj-methods}
\alias{KernelObj-methods}
\alias{.getKernelX,data.frame,KernelObj-method}
\alias{.kernel,KernelObj,ANY,ANY-method}
\alias{.kernel,KernelObj,missing,missing-method}
\alias{.kernel,KernelObj,missing,ANY-method}
\alias{.kernel,KernelObj,ANY,missing-method}
\alias{print,KernelObj-method}
\alias{show,KernelObj-method}
\alias{summary,KernelObj-method}
\title{Methods Available for Objects of Class \code{KernelObj}}
\usage{
\S4method{.getKernelX}{data.frame,KernelObj}(data, object)

\S4method{.kernel}{KernelObj,ANY,ANY}(object, x1, x2, ...)

\S4method{.kernel}{KernelObj,missing,missing}(object, x1, x2, ...)

\S4method{.kernel}{KernelObj,missing,ANY}(object, x1, x2, ...)

\S4method{.kernel}{KernelObj,ANY,missing}(object, x1, x2, ...)

\S4method{print}{KernelObj}(x, ...)

\S4method{show}{KernelObj}(object)

\S4method{summary}{KernelObj}(object, ...)
}
\description{
Methods Available for Objects of Class \code{KernelObj}

\code{.getKernelX}
  not allowed.

\code{.kernel}
  not allowed.

\code{print}
  prints kernel model. Includes "Kernel" as header.

\code{show}
  displays kernel model. Includes "Kernel" as header.

\code{summary}
  not allowed.
}
\keyword{internal}
