% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O_Learning.R
\docType{class}
\name{Learning-class}
\alias{Learning-class}
\title{Class \code{Learning}}
\description{
Class \code{Learning} contains results for a learning analysis with one 
  regime.
}
\section{Slots}{

\describe{
\item{\code{txInfo}}{Feasible tx information}

\item{\code{propen}}{Propensity regression analysis}

\item{\code{outcome}}{Outcome regression analysis}

\item{\code{optim}}{Optimization analysis}
}}

\keyword{internal}
