% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O_Learning.R
\name{Learning-methods}
\alias{Learning-methods}
\alias{Call,Learning-method}
\alias{cvInfo,Learning-method}
\alias{coef,Learning-method}
\alias{estimator,Learning-method}
\alias{fitObject,Learning-method}
\alias{optimObj,Learning-method}
\alias{optTx,Learning,data.frame-method}
\alias{optTx,Learning,missing-method}
\alias{outcome,Learning-method}
\alias{plot,Learning,ANY-method}
\alias{print,Learning-method}
\alias{propen,Learning-method}
\alias{regimeCoef,Learning-method}
\alias{show,Learning-method}
\alias{summary,Learning-method}
\title{Methods Available for Objects of Class \code{Learning}}
\usage{
\S4method{Call}{Learning}(name, ...)

\S4method{cvInfo}{Learning}(object, ...)

\S4method{coef}{Learning}(object, ...)

\S4method{estimator}{Learning}(x, ...)

\S4method{fitObject}{Learning}(object, ...)

\S4method{optimObj}{Learning}(object, ...)

\S4method{optTx}{Learning,data.frame}(x, newdata)

\S4method{optTx}{Learning,missing}(x, newdata, ...)

\S4method{outcome}{Learning}(object, ...)

\S4method{plot}{Learning,ANY}(x, suppress = FALSE, ...)

\S4method{print}{Learning}(x, ...)

\S4method{propen}{Learning}(object, ...)

\S4method{regimeCoef}{Learning}(object, ...)

\S4method{show}{Learning}(object)

\S4method{summary}{Learning}(object, ...)
}
\description{
Methods Available for Objects of Class \code{Learning}
}
\keyword{internal}
