% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O_LearningMulti.R
\docType{class}
\name{LearningMulti-class}
\alias{LearningMulti-class}
\title{Class \code{LearningMulti}}
\description{
Class \code{LearningMulti} contains results for a learning analysis 
  with multiple regimes.
}
\section{Slots}{

\describe{
\item{\code{optimResult}}{ANY containing a list of OptimStep results}
}}

\keyword{internal}
