% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O_LearningMulti.R
\docType{methods}
\name{LearningMulti-methods}
\alias{LearningMulti-methods}
\alias{Call,LearningMulti-method}
\alias{cvInfo,LearningMulti-method}
\alias{coef,LearningMulti-method}
\alias{estimator,LearningMulti-method}
\alias{fitObject,LearningMulti-method}
\alias{optimObj,LearningMulti-method}
\alias{optTx,LearningMulti,data.frame-method}
\alias{optTx,LearningMulti,missing-method}
\alias{outcome,LearningMulti-method}
\alias{plot,LearningMulti,ANY-method}
\alias{print,LearningMulti-method}
\alias{propen,LearningMulti-method}
\alias{regimeCoef,LearningMulti-method}
\alias{show,LearningMulti-method}
\alias{summary,LearningMulti-method}
\title{Methods Available for Objects of Class \code{LearningMulti}}
\usage{
\S4method{Call}{LearningMulti}(name, ...)

\S4method{cvInfo}{LearningMulti}(object, ...)

\S4method{coef}{LearningMulti}(object, ...)

\S4method{estimator}{LearningMulti}(x, ...)

\S4method{fitObject}{LearningMulti}(object, ...)

\S4method{optimObj}{LearningMulti}(object, ...)

\S4method{optTx}{LearningMulti,data.frame}(x, newdata, ...)

\S4method{optTx}{LearningMulti,missing}(x, newdata, ...)

\S4method{outcome}{LearningMulti}(object, ...)

\S4method{plot}{LearningMulti,ANY}(x, suppress = FALSE, ...)

\S4method{print}{LearningMulti}(x, ...)

\S4method{propen}{LearningMulti}(object, ...)

\S4method{regimeCoef}{LearningMulti}(object, ...)

\S4method{show}{LearningMulti}(object)

\S4method{summary}{LearningMulti}(object, ...)
}
\description{
Methods Available for Objects of Class \code{LearningMulti}
}
\keyword{internal}
