% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O_LearningObject.R
\docType{class}
\name{LearningObject-class}
\alias{LearningObject-class}
\title{Class \code{LearningObject}}
\description{
Class \code{LearningObject} contains stores parameters required for 
 weighted learning optimization step
}
\section{Slots}{

\describe{
\item{\code{mu}}{Matrix of outcome regression}

\item{\code{txVec}}{Vector of treatment coded as -1/1}

\item{\code{invPi}}{Vector of inverse propensity for treatment received}

\item{\code{response}}{Vector of the response}
}}

\keyword{internal}
