% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O_LearningObject.R
\name{LearningObject-methods}
\alias{LearningObject-methods}
\alias{.subsetObject,LearningObject-method}
\alias{.objFn,numeric,LearningObject,Kernel-method}
\alias{.dobjFn,numeric,LearningObject,Kernel-method}
\alias{.valueFunc,LearningObject-method}
\alias{.createLearningObject}
\title{Methods Available for Objects of Class \code{LearningObject}}
\usage{
\S4method{.subsetObject}{LearningObject}(methodObject, subset)

\S4method{.objFn}{numeric,LearningObject,Kernel}(par, methodObject, kernel, lambda, ...)

\S4method{.dobjFn}{numeric,LearningObject,Kernel}(par, methodObject, kernel, lambda, ...)

\S4method{.valueFunc}{LearningObject}(methodObject, optTx, ...)

.createLearningObject(
  kernel,
  surrogate,
  mu,
  txVec,
  response,
  prWgt,
  guess = NULL,
  ...
)
}
\arguments{
\item{kernel}{Kernel object}

\item{surrogate}{Surrogate object indicating surrogate loss-function}

\item{mu}{Matrix of predicted outcome on binary tx coding}

\item{txVec}{Tx vector coded as -1/1}

\item{response}{vector of response}

\item{prWgt}{propensity wgt for tx received}

\item{guess}{Vector of estimated regime parameters}
}
\value{
A LearningObject object
}
\description{
Methods Available for Objects of Class \code{LearningObject}

Create LearningObject
}
\keyword{internal}
