% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_LinearKernel.R
\docType{class}
\name{LinearKernel-class}
\alias{LinearKernel-class}
\title{Class \code{LinearKernel}}
\description{
Class \code{LinearKernel} holds information regarding decision function 
  when kernel is linear
}
\keyword{internal}
