% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_LinearKernel.R
\docType{methods}
\name{LinearKernel-methods}
\alias{LinearKernel-methods}
\alias{.kernelNumPars,LinearKernel-method}
\alias{.kernel,LinearKernel,matrix,matrix-method}
\alias{print,LinearKernel-method}
\alias{show,LinearKernel-method}
\alias{summary,LinearKernel-method}
\title{Methods Available for Objects of Class \code{LinearKernel}}
\usage{
\S4method{.kernelNumPars}{LinearKernel}(object, ...)

\S4method{.kernel}{LinearKernel,matrix,matrix}(object, x1, x2, ...)

\S4method{print}{LinearKernel}(x, ...)

\S4method{show}{LinearKernel}(object)

\S4method{summary}{LinearKernel}(object, ...)
}
\description{
Methods Available for Objects of Class \code{LinearKernel}
}
\keyword{internal}
