% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_List.R
\docType{class}
\name{List-class}
\alias{List-class}
\title{Class \code{List}}
\description{
Class \code{List} mimics a \code{list}.
}
\section{Slots}{

\describe{
\item{\code{names}}{Character vector of names for elements}
}}

\keyword{internal}
