% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_LogitSurrogate.R
\name{LogitSurrogate-methods}
\alias{LogitSurrogate-methods}
\alias{.phiFunc,LogitSurrogate-method}
\alias{.dPhiFunc,LogitSurrogate-method}
\title{Methods Available for Objects of Class \code{LogitSurrogate}}
\usage{
\S4method{.phiFunc}{LogitSurrogate}(surrogate, u)

\S4method{.dPhiFunc}{LogitSurrogate}(surrogate, u, du)
}
\description{
Methods Available for Objects of Class \code{LogitSurrogate}

\code{.phiFunc}
  calculates logistic surrogate loss-function

\code{.dphiFunc}
  calculates derivative of logistic surrogate loss-function
}
\keyword{internal}
