% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_MethodObject.R
\docType{class}
\name{MethodObject-class}
\alias{MethodObject-class}
\title{Class \code{MethodObject}}
\description{
Class \code{MethodObject} stores parameters required for optimization step
}
\section{Slots}{

\describe{
\item{\code{x}}{Matrix of covariates for kernel}

\item{\code{surrogate}}{The Surrogate for the loss-function}

\item{\code{pars}}{Vector of regime parameters}

\item{\code{kernel}}{The Kernel defining the decision function}
}}

\keyword{internal}
