% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_MethodObject.R
\name{MethodObject-methods}
\alias{MethodObject-methods}
\alias{.subsetObject,MethodObject-method}
\alias{.objFn,numeric,MethodObject,Kernel-method}
\alias{.dobjFn,numeric,MethodObject,Kernel-method}
\alias{.valueFunc,MethodObject-method}
\alias{.createMethodObject}
\title{Methods Available for Objects of Class \code{MethodObject}}
\usage{
\S4method{.subsetObject}{MethodObject}(methodObject, subset)

\S4method{.objFn}{numeric,MethodObject,Kernel}(par, methodObject, kernel, lambda, ...)

\S4method{.dobjFn}{numeric,MethodObject,Kernel}(par, methodObject, kernel, lambda, ...)

\S4method{.valueFunc}{MethodObject}(methodObject, optTx, ...)

.createMethodObject(kernel, surrogate, guess = NULL, ...)
}
\arguments{
\item{kernel}{Kernel object}

\item{surrogate}{Surrogate object indicating surrogate loss-function}

\item{guess}{Vector of estimated regime parameters}
}
\value{
A MethodObject object
}
\description{
Methods Available for Objects of Class \code{MethodObject}

Create method object
}
\keyword{internal}
