\name{fittedMain}
\alias{fittedMain}
\alias{fittedMain,DynTxRegime-method}

\title{ Extract Fitted Main Effects}
\description{
Extract fitted main effects for IQ-Learning algorithm.
}

\usage{ \S4method{fittedMain}{DynTxRegime}(object, ...)  }

\arguments{
  \item{object }{an object of class DynTxRegime.}

  \item{\dots }{ignored.}

}

\value{
For object returned by IQ-Learning functions, the fitted main effects.
For all others, NULL
}

\author{
  Shannon T. Holloway <sthollow@ncsu.edu>
}


