\name{DTRstep}
\docType{methods}
\alias{DTRstep}

\title{ Identify Statistical Method of DynTxRegime Object}
\description{
 Retrieves a description of the method used to create the object.
}

\usage{
DTRstep(object)
}

\arguments{
\item{object}{An object of class \code{"DynTxRegime."}}
}

\author{
  Shannon T. Holloway <sthollow@ncsu.edu>
}

