\name{summary}
\docType{methods}
\alias{summary}
\alias{summary,DynTxRegime-method}

\title{ Summary Results}
\description{
Retrieve summary results for a DynTxRegime method.
}
\usage{
summary(object,...)
}

\arguments{
\item{object}{An object of class \code{"DynTxRegime"}.}
\item{...}{Additional arguments affecting the summary produced.}q
}

