% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_ModelObj_SubsetList.R
\docType{class}
\name{ModelObj_SubsetList-class}
\alias{ModelObj_SubsetList-class}
\alias{initialize,ModelObj_SubsetList-method}
\title{Class ModelObj_SubsetList}
\usage{
\S4method{initialize}{ModelObj_SubsetList}(.Object, ...)
}
\description{
Class \code{ModelObj_SubsetList} represents a \code{List} for subset
modelObj.
}
\keyword{internal}
