\name{ModelObj_SubsetList_DecisionPointList-class}
\docType{class}
\alias{ModelObj_SubsetList_DecisionPointList-class}

\title{Class \code{"ModelObj_SubsetList_DecisionPointList"}}

\description{
List of objects of class \code{"ModelObj_SubsetList"}
for multiple-decision-point methods.
}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ModelObj_SubsetList_DecisionPointList", ...)}.
These objects are for convenience in package development
and should not be created by users.
}

\section{Extends}{
Class \code{"\link{DecisionPointList-class}"}, directly.
Classes \code{"\link{List-class}"} and \code{"\link{MultipleDecisionPoint-class}"}, by class "DecisionPointList", distance 2.
}

\section{Methods}{
No methods defined with class \code{"ModelObj_SubsetList_DecisionPointList"} in the signature.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("ModelObj_SubsetList_DecisionPointList")
}

\keyword{internal}
