% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_ModelObjSubset.R
\name{ModelObjSubset-methods}
\alias{ModelObjSubset-methods}
\alias{.getDecisionPoint,ModelObjSubset-method}
\alias{.getSubset,ModelObjSubset-method}
\title{Methods Available for Objects of Class \code{ModelObjSubset}}
\usage{
\S4method{.getDecisionPoint}{ModelObjSubset}(object)

\S4method{.getSubset}{ModelObjSubset}(object)
}
\description{
Methods Available for Objects of Class \code{ModelObjSubset}

\code{.getDecisionPoint(object)}
  retrieves the decision point to which object pertains

\code{.getSubset(object)}
  retrieves the subset to which object pertains
}
\keyword{internal}
