\name{MultipleDecisionPoint-class}
\docType{class}
\alias{MultipleDecisionPoint-class}

\title{Class \code{"MultipleDecisionPoint"}}

\description{
A class to indicate that an object is for multiple decision points.
}

\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}

\section{Methods}{
  No methods exist for this class.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("MultipleDecisionPoint")
}

\keyword{internal}
