% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_MultiRadialKernel.R
\docType{class}
\name{MultiRadialKernel-class}
\alias{MultiRadialKernel-class}
\title{Class \code{MultiRadialKernel}}
\description{
Class \code{MultiRadialKernel} holds information regarding decision function 
  when kernel is radial and multiple kernel parameters
}
\keyword{internal}
