% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_MultiRadialKernel.R
\docType{methods}
\name{MultiRadialKernel-methods}
\alias{MultiRadialKernel-methods}
\alias{.kernel,MultiRadialKernel,matrix,matrix-method}
\alias{print,MultiRadialKernel-method}
\alias{show,MultiRadialKernel-method}
\alias{summary,MultiRadialKernel-method}
\title{Methods Available for Objects of Class \code{MultiRadialKernel}}
\usage{
\S4method{.kernel}{MultiRadialKernel,matrix,matrix}(object, x1, x2, ...)

\S4method{print}{MultiRadialKernel}(x, ...)

\S4method{show}{MultiRadialKernel}(object)

\S4method{summary}{MultiRadialKernel}(object, ...)
}
\description{
Methods Available for Objects of Class \code{MultiRadialKernel}

\code{.kernel}
  not allowed.

\code{print}
  not allowed.

\code{show}
  not allowed.

\code{summary}
  not allowed.
}
\keyword{internal}
