% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_class_BOWL.R
\docType{methods}
\name{newBOWL}
\alias{newBOWL}
\alias{.newBOWL,BOWL-method}
\title{Create a BOWL Object for First Step of BOWL Algorithm}
\usage{
\S4method{.newBOWL}{`NULL`}(BOWLObj, moPropen, fSet, data, response,
  txName, lambdas, cvFolds, kernel, surrogate, suppress, guess, ...)

\S4method{.newBOWL}{BOWL}(BOWLObj, moPropen, fSet, data, response, txName,
  lambdas, cvFolds, kernel, surrogate, suppress, guess, ...)
}
\arguments{
\item{BOWLObj}{NULL or an object returned from a previous step}

\item{moPropen}{modelObj or modelObjSubset for propensity modeling}

\item{fSet}{optional function defining subsets for modeling}

\item{data}{data.frame of covariates}

\item{response}{response}

\item{txName}{treatment variable column header in data}

\item{lambdas}{tuning parameter(s)}

\item{cvFolds}{number of cross-validation folds}

\item{kernel}{Kernel object}

\item{surrogate}{Surrogate object}

\item{suppress}{T/F indicating if prints to screen are to be executed}

\item{guess}{Starting values for optimization}
}
\description{
Create a BOWL Object for First Step of BOWL Algorithm
}
\keyword{internal}
