% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_class_BOWLBasic.R
\name{newBOWLStep}
\alias{newBOWLStep}
\alias{.newBOWLStep}
\title{Create a BOWL Object}
\usage{
.newBOWLStep(
  moPropen,
  fSet,
  data,
  response,
  txName,
  lambdas,
  cvFolds,
  kernel,
  surrogate,
  suppress,
  guess,
  prodPi,
  index,
  ...
)
}
\arguments{
\item{moPropen}{model object for propensity}

\item{fSet}{function specifying subsets or NULL}

\item{data}{data.frame of covariates and tx}

\item{response}{vector of responses}

\item{txName}{character indicating tx column in data}

\item{lambdas}{vector of tuning parameters}

\item{cvFolds}{number of cross-validation folds or NULL}

\item{kernel}{Kernel object}

\item{surrogate}{Surrogate object}

\item{guess}{vector of starting value for regime parameterse}

\item{prodPi}{vector of previous step propensity weights}

\item{index}{vector indicating previous compliance with regime}

\item{...}{additional inputs sent to optimization method}
}
\value{
BOWLBasic object
}
\description{
Create a BOWL Object
}
\keyword{internal}
