% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_ClassificationFit.R,
%   R/I_ClassificationFit_SubsetList.R, R/I_ClassificationFit_fSet.R
\docType{methods}
\name{newClassificationFit}
\alias{newClassificationFit}
\alias{.newClassificationFit,modelObj,TxInfoNoSubsets-method}
\alias{.newClassificationFit,ModelObj_SubsetList,TxInfoWithSubsets-method}
\alias{.newClassificationFit,modelObj,TxInfoWithSubsets-method}
\title{Complete a Classification Regression Step}
\usage{
.newClassificationFit(moClass, txObj, ...)

\S4method{.newClassificationFit}{modelObj,TxInfoNoSubsets}(moClass, txObj, response, data, suppress, ...)

\S4method{.newClassificationFit}{ModelObj_SubsetList,TxInfoWithSubsets}(moClass, data, response, txObj, suppress, ...)

\S4method{.newClassificationFit}{modelObj,TxInfoWithSubsets}(moClass, txObj, response, data, suppress, ...)
}
\arguments{
\item{moClass}{modeling object(s) defining the classification regression}

\item{txObj}{TxObj defining the tx feasible sets}

\item{...}{additional arguments. Ignored.}

\item{data}{data.frame of covariates and tx received}

\item{suppress}{logical indicating user's screen printing preference}

\item{contrast}{vector of contrasts}
}
\description{
Methods dispatch appropriate typed fit methods based on the modeling
  object specified by the user and the feasible tx definitions.
  The value object returned depends on the underlying typed fit method.
}
\keyword{internal}
