% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVInfo.R, R/N_CVInfoLambda.R,
%   R/N_CVInfokParam.R, R/N_CVInfo2Par.R
\name{.newCVInfo}
\alias{.newCVInfo}
\alias{.newCVInfo,ANY,ANY-method}
\alias{.newCVInfo,numeric,Kernel-method}
\alias{.newCVInfo,numeric,MultiRadialKernel-method}
\alias{.newCVInfo,array,Kernel-method}
\alias{.newCVInfo,array,MultiRadialKernel-method}
\title{Create a CVInfo Object}
\usage{
.newCVInfo(lambdas, kernel, ...)

\S4method{.newCVInfo}{ANY,ANY}(lambdas, kernel, ...)

\S4method{.newCVInfo}{numeric,Kernel}(lambdas, kernel, methodObject, cvObject, suppress, ...)

\S4method{.newCVInfo}{numeric,MultiRadialKernel}(lambdas, kernel, methodObject, cvObject, suppress, ...)

\S4method{.newCVInfo}{array,Kernel}(lambdas, kernel, methodObject, cvObject, suppress, ...)

\S4method{.newCVInfo}{array,MultiRadialKernel}(lambdas, kernel, methodObject, cvObject, suppress, ...)

\S4method{.newCVInfo}{numeric,MultiRadialKernel}(lambdas, kernel, methodObject, cvObject, suppress, ...)

\S4method{.newCVInfo}{array,MultiRadialKernel}(lambdas, kernel, methodObject, cvObject, suppress, ...)
}
\arguments{
\item{lambdas}{tuning parameters}

\item{kernel}{kernel object}
}
\description{
Dispatch appropriate cross-validation procedure.
}
\keyword{internal}
