% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVInfoObj.R
\name{newCVInfoObj}
\alias{newCVInfoObj}
\alias{.newCVInfoObj,ANY,Kernel-method}
\title{Create a New CVInfoObj Object}
\usage{
.newCVInfoObj(lambdas, kernel, ...)

\S4method{.newCVInfoObj}{ANY,Kernel}(lambdas, kernel, methodObject, cvObject, suppress, ...)
}
\arguments{
\item{lambdas}{Tuning parameters to be considered}

\item{kernel}{Kernel (w/kernel parameters) to be considered}

\item{...}{Additional arguments as needed}

\item{methodObject}{Object parameters for weighted learning method}

\item{suppress}{T/F indicating if screen prints are generated}

\item{cvFolds}{Number of cross-validation folds}
}
\description{
Call newCVInfo and stores result in @cvInfo
}
\keyword{internal}
