% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVBasic.R
\name{newCVStep}
\alias{newCVStep}
\alias{.newCVStep}
\title{An n-Fold Cross Validation Step}
\usage{
.newCVStep(cvObject, methodObject, lambda, suppress, ...)
}
\arguments{
\item{cvObject}{Information regarding folds and treatment groups}

\item{methodObject}{Information needed for method specific objective function}

\item{lambda}{numeric A single tuning parameter value}

\item{suppress}{integer indicating printing preference}

\item{...}{additional inputs.}
}
\value{
The average value across all successfully trained folds
}
\description{
An n-Fold Cross Validation Step
}
\keyword{internal}
