% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_class_EARL.R
\name{newEARL}
\alias{newEARL}
\alias{.newEARL}
\title{Complete an EARL Analysis}
\usage{
.newEARL(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  lambdas,
  cvFolds,
  surrogate,
  iter,
  guess,
  kernel,
  fSet,
  suppress,
  ...
)
}
\arguments{
\item{moPropen}{modelObj for propensity modeling}

\item{moMain}{modelObj for main effects of outcome model}

\item{moCont}{modelObj for contrasts of outcome model}

\item{data}{data.frame of covariates}

\item{response}{Vector of responses}

\item{txName}{Tx variable column header in data}

\item{lambdas}{Tuning parameter(s)}

\item{cvFolds}{Number of cross-validation folds}

\item{surrogate}{Surrogate object}

\item{iter}{Maximum iterations for outcome regression}

\item{guess}{optional numeric vector providing starting values for
optimization methods}

\item{kernel}{Kernel object or SubsetList}

\item{fSet}{NULL or function defining subset rules}

\item{suppress}{T/F indicating if prints to screen are executed}

\item{...}{Additional inputs for optimization}

\item{txVec}{Tx vector recast as +/- 1}
}
\value{
An EARL object
}
\description{
Complete an EARL Analysis
}
\keyword{internal}
