% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnFS_C.R
\name{newIQLearnFS_C}
\alias{newIQLearnFS_C}
\alias{.newIQLearnFS_C,modelObj,modelObj,IQLearnSS-method}
\alias{.newIQLearnFS_C,modelObj,NULL,IQLearnSS-method}
\alias{.newIQLearnFS_C,NULL,modelObj,IQLearnSS-method}
\title{Complete First Stage Analysis of Contrasts for Interactive Q-Learning Algorithm}
\usage{
.newIQLearnFS_C(moMain, moCont, response, ...)

\S4method{.newIQLearnFS_C}{modelObj,modelObj,IQLearnSS}(moMain, moCont, response, data, txName, iter, suppress)

\S4method{.newIQLearnFS_C}{modelObj,`NULL`,IQLearnSS}(moMain, moCont, response, data, txName, iter, suppress)

\S4method{.newIQLearnFS_C}{`NULL`,modelObj,IQLearnSS}(moMain, moCont, response, data, txName, iter, suppress)
}
\description{
Performs regression on the fitted contrasts of the second stage regression.
}
\keyword{internal}
