% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnFS_VHet.R
\docType{methods}
\name{newIQLearnFS_VHet}
\alias{newIQLearnFS_VHet}
\alias{.newIQLearnFS_VHet,IQLearnFS_C,modelObj,modelObj-method}
\alias{.newIQLearnFS_VHet,IQLearnFS_C,modelObj,NULL-method}
\alias{.newIQLearnFS_VHet,IQLearnFS_C,NULL,modelObj-method}
\title{Complete First Stage Analysis of Residuals for Interactive Q-Learning Algorithm}
\usage{
.newIQLearnFS_VHet(object, moMain, moCont, ...)

\S4method{.newIQLearnFS_VHet}{IQLearnFS_C,modelObj,modelObj}(object,
  moMain, moCont, data, iter, suppress)

\S4method{.newIQLearnFS_VHet}{IQLearnFS_C,modelObj,`NULL`}(object, moMain,
  moCont, data, iter, suppress)

\S4method{.newIQLearnFS_VHet}{IQLearnFS_C,`NULL`,modelObj}(object, moMain,
  moCont, data, iter, suppress)
}
\description{
Performs log-linear regression on the residuals.
}
\keyword{internal}
