% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnSS.R
\docType{methods}
\name{newIQLearnSS}
\alias{newIQLearnSS}
\alias{.newIQLearnSS,modelObj,modelObj-method}
\alias{.newIQLearnSS,modelObj,NULL-method}
\alias{.newIQLearnSS,NULL,modelObj-method}
\title{Complete Second Stage Analysis of Interactive Q-Learning Algorithm}
\usage{
.newIQLearnSS(moMain, moCont, ...)

\S4method{.newIQLearnSS}{modelObj,modelObj}(moMain, moCont, data, response,
  txName, iter, suppress)

\S4method{.newIQLearnSS}{modelObj,`NULL`}(moMain, moCont, data, response,
  txName, iter, suppress)

\S4method{.newIQLearnSS}{`NULL`,modelObj}(moMain, moCont, data, response,
  txName, iter, suppress)
}
\description{
Performs the regression of the outcome.
}
\keyword{internal}
