% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_newModel.R
\name{.newModel}
\alias{.newModel}
\alias{.newModel,modelObj,modelObj-method}
\alias{.newModel,modelObj,NULL-method}
\alias{.newModel,NULL,modelObj-method}
\title{Combine model object models}
\usage{
.newModel(moMain, moCont, ...)

\S4method{.newModel}{modelObj,modelObj}(moMain, moCont, txName, suppress)

\S4method{.newModel}{modelObj,`NULL`}(moMain, moCont, txName, suppress)

\S4method{.newModel}{`NULL`,modelObj}(moMain, moCont, txName, suppress)
}
\description{
Combines moMain and moCont into a single modeling object.
}
\keyword{internal}
