% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_newModelObjSubset.R
\name{newModelObjSubset}
\alias{newModelObjSubset}
\alias{.newModelObjSubset}
\title{Create Internal Model Objects for Subsets of Data}
\usage{
.newModelObjSubset(object)
}
\arguments{
\item{object}{A list of modelObj or ModelObjSubset}
}
\value{
An object of class \code{ModelObj_SubsetList} if a single decision 
  point or an object of class \code{ModelObj_DecisionPointList} if multiple 
  decision points.
}
\description{
Create Internal Model Objects for Subsets of Data
}
\keyword{internal}
