% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_OptimBasic.R, R/M_OptimKernel.R
\name{.newOptim}
\alias{.newOptim}
\alias{.newOptim,LinearKernel-method}
\alias{.newOptim,Kernel-method}
\title{Complete an Optimization Step}
\usage{
.newOptim(kernel, ...)

\S4method{.newOptim}{LinearKernel}(kernel, lambda, methodObject, suppress, ...)

\S4method{.newOptim}{Kernel}(kernel, lambda, methodObject, suppress, ...)
}
\description{
Dispatches appropriate methods to optimize an object function.
}
\keyword{internal}
