% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_class_OptimalClass.R
\name{newOptimalClass}
\alias{newOptimalClass}
\alias{.newOptimalClass,vector-method}
\alias{.newOptimalClass,OptimalClass-method}
\title{Estimate the Optimal Treatment and Value Using Classification}
\usage{
.newOptimalClass(response, ...)

\S4method{.newOptimalClass}{vector}(
  moPropen,
  moMain,
  moCont,
  moClass,
  data,
  response,
  txName,
  iter,
  fSet,
  suppress,
  ...
)

\S4method{.newOptimalClass}{OptimalClass}(
  moPropen,
  moMain,
  moCont,
  moClass,
  data,
  response,
  txName,
  iter,
  fSet,
  suppress,
  ...
)
}
\description{
Method dispatches the appropriate function to obtain estimates for the
  optimal treatment and value using classification.
}
\keyword{internal}
