% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_class_OptimalSeq.R,
%   R/H_class_OptimalSeqCoarsened.R, R/H_class_OptimalSeqMissing.R
\name{newOptimalSeq}
\alias{newOptimalSeq}
\alias{.newOptimalSeq,ModelObj_DecisionPointList,ModelObj_DecisionPointList,ModelObj_DecisionPointList,list-method}
\alias{.newOptimalSeq,ModelObj_DecisionPointList,ModelObj_DecisionPointList,NULL,list-method}
\alias{.newOptimalSeq,ModelObj_DecisionPointList,NULL,ModelObj_DecisionPointList,list-method}
\alias{.newOptimalSeq,ModelObj_DecisionPointList,NULL,NULL,list-method}
\alias{.newOptimalSeq,ModelObj_DecisionPointList,ModelObj_DecisionPointList,ModelObj_DecisionPointList,NULL-method}
\alias{.newOptimalSeq,ModelObj_DecisionPointList,ModelObj_DecisionPointList,NULL,NULL-method}
\alias{.newOptimalSeq,ModelObj_DecisionPointList,NULL,ModelObj_DecisionPointList,NULL-method}
\alias{.newOptimalSeq,ModelObj_DecisionPointList,NULL,NULL,NULL-method}
\alias{.newOptimalSeq,modelObj,modelObj,modelObj,NULL-method}
\alias{.newOptimalSeq,modelObj,modelObj,NULL,NULL-method}
\alias{.newOptimalSeq,modelObj,NULL,modelObj,NULL-method}
\alias{.newOptimalSeq,modelObj,NULL,NULL,NULL-method}
\alias{.newOptimalSeq,modelObj,modelObj,modelObj,function-method}
\alias{.newOptimalSeq,modelObj,modelObj,NULL,function-method}
\alias{.newOptimalSeq,modelObj,NULL,modelObj,function-method}
\alias{.newOptimalSeq,modelObj,NULL,NULL,function-method}
\alias{.newOptimalSeq,ModelObj_SubsetList,modelObj,modelObj,function-method}
\alias{.newOptimalSeq,ModelObj_SubsetList,modelObj,NULL,function-method}
\alias{.newOptimalSeq,ModelObj_SubsetList,NULL,modelObj,function-method}
\alias{.newOptimalSeq,ModelObj_SubsetList,NULL,NULL,function-method}
\alias{.newOptimalSeq,ModelObj_SubsetList,ModelObj_SubsetList,ModelObj_SubsetList,function-method}
\alias{.newOptimalSeq,ModelObj_SubsetList,ModelObj_SubsetList,NULL,function-method}
\alias{.newOptimalSeq,ModelObj_SubsetList,NULL,ModelObj_SubsetList,function-method}
\alias{.newOptimalSeq,modelObj,ModelObj_SubsetList,ModelObj_SubsetList,function-method}
\alias{.newOptimalSeq,modelObj,ModelObj_SubsetList,NULL,function-method}
\alias{.newOptimalSeq,modelObj,NULL,ModelObj_SubsetList,function-method}
\title{Complete a the Coarsened/Missing Data Analysis}
\usage{
.newOptimalSeq(moPropen, moMain, moCont, fSet, ...)

\S4method{.newOptimalSeq}{ModelObj_DecisionPointList,ModelObj_DecisionPointList,ModelObj_DecisionPointList,list}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_DecisionPointList,ModelObj_DecisionPointList,`NULL`,list}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_DecisionPointList,`NULL`,ModelObj_DecisionPointList,list}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_DecisionPointList,`NULL`,`NULL`,list}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_DecisionPointList,ModelObj_DecisionPointList,ModelObj_DecisionPointList,`NULL`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_DecisionPointList,ModelObj_DecisionPointList,`NULL`,`NULL`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_DecisionPointList,`NULL`,ModelObj_DecisionPointList,`NULL`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_DecisionPointList,`NULL`,`NULL`,`NULL`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,modelObj,modelObj,`NULL`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,modelObj,`NULL`,`NULL`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,`NULL`,modelObj,`NULL`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,`NULL`,`NULL`,`NULL`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,modelObj,modelObj,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,modelObj,`NULL`,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,`NULL`,modelObj,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,`NULL`,`NULL`,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_SubsetList,modelObj,modelObj,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_SubsetList,modelObj,`NULL`,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_SubsetList,`NULL`,modelObj,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_SubsetList,`NULL`,`NULL`,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_SubsetList,ModelObj_SubsetList,ModelObj_SubsetList,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_SubsetList,ModelObj_SubsetList,`NULL`,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{ModelObj_SubsetList,`NULL`,ModelObj_SubsetList,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,ModelObj_SubsetList,ModelObj_SubsetList,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,ModelObj_SubsetList,`NULL`,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)

\S4method{.newOptimalSeq}{modelObj,`NULL`,ModelObj_SubsetList,`function`}(
  moPropen,
  moMain,
  moCont,
  data,
  response,
  txName,
  regimesObj,
  fSet,
  iter,
  suppress,
  argsList,
  ...
)
}
\arguments{
\item{moPropen}{model object(s) for propensity regression}

\item{moMain}{model object(s) for main effects of outcome regression}

\item{moCont}{model object(s) for contrasts of outcome regression}

\item{fSet}{function(s) defining feasible tx}

\item{...}{additional inputs.}
}
\description{
Dispatches appropriate coarsened or missing data perspective method.
}
\keyword{internal}
