% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_OptimObj.R, R/Q_class_.rwl.R
\name{newOptimObj}
\alias{newOptimObj}
\alias{.newOptimObj,ANY-method}
\alias{.newOptimObj,.rwl-method}
\title{Create an OptimObj Object}
\usage{
.newOptimObj(methodObject, kernel, ...)

\S4method{.newOptimObj}{ANY}(methodObject, lambda, suppress, ...)

\S4method{.newOptimObj}{.rwl}(methodObject, lambda, suppress, ...)
}
\arguments{
\item{methodObject}{object containing parameters needed by a weighted 
learning method}

\item{...}{additional inputs passed to optimization routine.}

\item{lambda}{tuning parameters}

\item{suppress}{integer indicating screen print preferences}
}
\description{
Call newOptim and stores result under common name
}
\keyword{internal}
