% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeNoFit.R, R/D_OutcomeSimpleFit.R,
%   R/D_OutcomeSimpleFit_fSet.R, R/D_OutcomeIterateFit.R,
%   R/D_OutcomeSimpleFit_SubsetList.R
\name{newOutcomeFit}
\alias{newOutcomeFit}
\alias{.newOutcomeFit,NULL,NULL,TxObj,NULL-method}
\alias{.newOutcomeFit,modelObj,modelObj,TxInfoNoSubsets,NULL-method}
\alias{.newOutcomeFit,modelObj,NULL,TxInfoNoSubsets,NULL-method}
\alias{.newOutcomeFit,NULL,modelObj,TxInfoNoSubsets,NULL-method}
\alias{.newOutcomeFit,modelObj,modelObj,TxInfoWithSubsets,NULL-method}
\alias{.newOutcomeFit,modelObj,NULL,TxInfoWithSubsets,NULL-method}
\alias{.newOutcomeFit,NULL,modelObj,TxInfoWithSubsets,NULL-method}
\alias{.newOutcomeFit,modelObj,modelObj,TxInfoWithSubsets,integer-method}
\alias{.newOutcomeFit,modelObj,modelObj,TxInfoNoSubsets,integer-method}
\alias{.newOutcomeFit,ModelObj_SubsetList,ModelObj_SubsetList,TxInfoWithSubsets,integer-method}
\alias{.newOutcomeFit,ModelObj_SubsetList,ModelObj_SubsetList,TxInfoWithSubsets,ANY-method}
\alias{.newOutcomeFit,ModelObj_SubsetList,NULL,TxInfoWithSubsets,ANY-method}
\alias{.newOutcomeFit,NULL,ModelObj_SubsetList,TxInfoWithSubsets,ANY-method}
\title{Perform an Outcome Regression Step}
\usage{
.newOutcomeFit(moMain, moCont, txObj, iter, ...)

\S4method{.newOutcomeFit}{`NULL`,`NULL`,TxObj,`NULL`}(moMain, moCont, txObj, iter, data, response, suppress)

\S4method{.newOutcomeFit}{modelObj,modelObj,TxInfoNoSubsets,`NULL`}(moMain, moCont, txObj, iter, data, response, suppress)

\S4method{.newOutcomeFit}{modelObj,`NULL`,TxInfoNoSubsets,`NULL`}(moMain, moCont, txObj, iter, data, response, suppress)

\S4method{.newOutcomeFit}{`NULL`,modelObj,TxInfoNoSubsets,`NULL`}(moMain, moCont, txObj, iter, data, response, suppress)

\S4method{.newOutcomeFit}{modelObj,modelObj,TxInfoWithSubsets,`NULL`}(moMain, moCont, txObj, iter, data, response, suppress)

\S4method{.newOutcomeFit}{modelObj,`NULL`,TxInfoWithSubsets,`NULL`}(moMain, moCont, txObj, iter, data, response, suppress)

\S4method{.newOutcomeFit}{`NULL`,modelObj,TxInfoWithSubsets,`NULL`}(moMain, moCont, txObj, iter, data, response, suppress)

\S4method{.newOutcomeFit}{modelObj,modelObj,TxInfoWithSubsets,integer}(moMain, moCont, response, txObj, data, iter, suppress)

\S4method{.newOutcomeFit}{modelObj,modelObj,TxInfoNoSubsets,integer}(moMain, moCont, response, txObj, data, iter, suppress)

\S4method{.newOutcomeFit}{ModelObj_SubsetList,ModelObj_SubsetList,TxInfoWithSubsets,integer}(moMain, moCont, response, txObj, data, iter, suppress)

\S4method{.newOutcomeFit}{ModelObj_SubsetList,ModelObj_SubsetList,TxInfoWithSubsets,ANY}(moMain, moCont, txObj, data, response, iter, suppress)

\S4method{.newOutcomeFit}{ModelObj_SubsetList,`NULL`,TxInfoWithSubsets,ANY}(moMain, moCont, txObj, data, response, iter, suppress)

\S4method{.newOutcomeFit}{`NULL`,ModelObj_SubsetList,TxInfoWithSubsets,ANY}(moMain, moCont, txObj, data, response, iter, suppress)
}
\arguments{
\item{moMain}{A modeling object for main effects or NULL}

\item{moCont}{A modeling object for contrasts or NULL}

\item{txObj}{A TxObj object}

\item{iter}{NULL or numeric}

\item{...}{Any optional additional input.}
}
\description{
Dispatch appropriate methods to perform outcome regression step.
}
\keyword{internal}
