% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeObj.R
\name{.newOutcomeObj}
\alias{.newOutcomeObj}
\alias{.newOutcomeObj,ANY,ANY,ANY-method}
\alias{.newOutcomeObj,ModelObj_DecisionPointList,ModelObj_DecisionPointList,TxInfoList-method}
\alias{.newOutcomeObj,ModelObj_DecisionPointList,ANY,TxInfoList-method}
\alias{.newOutcomeObj,NULL,ModelObj_DecisionPointList,TxInfoList-method}
\title{Create a new \code{OutcomeObj} object}
\usage{
.newOutcomeObj(moMain, moCont, txObj, iter, ...)

\S4method{.newOutcomeObj}{ANY,ANY,ANY}(moMain, moCont, txObj, data, response, iter, suppress)

\S4method{.newOutcomeObj}{ModelObj_DecisionPointList,ModelObj_DecisionPointList,TxInfoList}(moMain, moCont, txObj, data, response, iter, suppress, ...)

\S4method{.newOutcomeObj}{ModelObj_DecisionPointList,ANY,TxInfoList}(moMain, moCont, txObj, data, response, iter, suppress, ...)

\S4method{.newOutcomeObj}{`NULL`,ModelObj_DecisionPointList,TxInfoList}(moMain, moCont, txObj, data, response, iter, suppress, ...)
}
\arguments{
\item{moMain}{A modeling object for main effects}

\item{moCont}{A modeling object for contrasts}

\item{txObj}{A TxObj object}

\item{iter}{NULL or integer}

\item{...}{Any optional additional input.}
}
\description{
Calls newOutcomeFit and stores in @outcome.
}
\keyword{internal}
