% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_class_OWL.R
\name{newOWL}
\alias{newOWL}
\alias{.newOWL}
\title{Complete an OWL Analysis}
\usage{
.newOWL(
  moPropen,
  data,
  response,
  txName,
  lambdas,
  cvFolds,
  kernel,
  fSet,
  surrogate,
  suppress,
  guess,
  ...
)
}
\arguments{
\item{moPropen}{modelObj for propensity modeling}

\item{data}{data.frame of covariates}

\item{response}{Vector of responses}

\item{txName}{Tx variable column header in data}

\item{lambdas}{Tuning parameter(s)}

\item{cvFolds}{Number of cross-validation folds}

\item{kernel}{Kernel object or SubsetList}

\item{fSet}{NULL or function defining subset rules}

\item{surrogate}{Surrogate object}

\item{suppress}{T/F indicating if prints to screen are executed}

\item{guess}{optional numeric vector providing starting values for
optimization methods}

\item{...}{Additional inputs for optimization}

\item{txVec}{Tx vector recast as +/- 1}
}
\value{
An OWL object
}
\description{
Complete an OWL Analysis
}
\keyword{internal}
