% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityFit.R, R/F_PropensityFit_fSet.R,
%   R/F_PropensityFit_SubsetList.R
\docType{methods}
\name{newPropensityFit}
\alias{newPropensityFit}
\alias{.newPropensityFit,modelObj,TxInfoNoSubsets-method}
\alias{.newPropensityFit,modelObj,TxInfoWithSubsets-method}
\alias{.newPropensityFit,ModelObj_SubsetList,TxInfoWithSubsets-method}
\title{Complete a Propensity Regression Step}
\usage{
.newPropensityFit(moPropen, txObj, ...)

\S4method{.newPropensityFit}{modelObj,TxInfoNoSubsets}(moPropen, txObj,
  data, suppress)

\S4method{.newPropensityFit}{modelObj,TxInfoWithSubsets}(moPropen, txObj,
  data, suppress)


  \S4method{.newPropensityFit}{ModelObj_SubsetList,TxInfoWithSubsets}(moPropen,
  txObj, data, suppress)
}
\arguments{
\item{moPropen}{A modeling object}

\item{txObj}{A TxObj object}

\item{...}{Any optional additional input.}
}
\description{
Dispatches appropriate method for completing propensity regressions.
}
\keyword{internal}
