% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityObj.R
\name{newPropensityObj}
\alias{newPropensityObj}
\alias{.newPropensityObj,ANY,ANY-method}
\alias{.newPropensityObj,ModelObj_DecisionPointList,TxInfoList-method}
\title{Create a new \code{PropensityObj} object}
\usage{
.newPropensityObj(moPropen, txObj, data, suppress, ...)

\S4method{.newPropensityObj}{ANY,ANY}(moPropen, txObj, data, suppress)

\S4method{.newPropensityObj}{ModelObj_DecisionPointList,TxInfoList}(moPropen, txObj, data, suppress)
}
\arguments{
\item{moPropen}{A modeling object}

\item{txObj}{A TxObj object}

\item{...}{Any optional additional input.}
}
\description{
Calls newPropensityFit and stores result in @propen.
}
\keyword{internal}
