% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_QLearn.R
\name{.newQLearn}
\alias{.newQLearn}
\alias{.newQLearn,vector-method}
\alias{.newQLearn,QLearn-method}
\title{Perform a Step of the Q-Learning Algorithm}
\usage{
.newQLearn(response, ...)

\S4method{.newQLearn}{vector}(moMain, moCont, fSet, response, data, txName, iter, suppress)

\S4method{.newQLearn}{QLearn}(moMain, moCont, fSet, response, data, txName, iter, suppress)
}
\arguments{
\item{response}{a vector or the value object returned by a prior call to 
qlearn()}

\item{moMain}{modeling object specifying the main effects component of the
outcome model}

\item{moCont}{modeling object specifying the contrasts component of the
outcome model}

\item{fSet}{function defining the feasible tx subsets}

\item{data}{data.frame of covariates and tx received}

\item{txName}{character name of tx variable in data}

\item{iter}{the maximum number of iterations in the iterative algorithm}

\item{suppress}{logical indicating user's screen printing preference}
}
\value{
an object of class QLearn.
}
\description{
Method performs all necessary regression and predictions steps for a
  single step of the Q-learning algorithm.
}
\keyword{internal}
