% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_Regime.R
\docType{methods}
\name{.newRegime}
\alias{.newRegime}
\alias{.newRegime,function-method}
\title{Create a new \code{Regime} object}
\usage{
.newRegime(object)

\S4method{.newRegime}{`function`}(object)
}
\arguments{
\item{object}{A function defining the treatment regime}
}
\description{
Create a new \code{Regime} object
}
\keyword{internal}
