% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_RegimeObj.R
\name{.newRegimeObj}
\alias{.newRegimeObj}
\alias{.newRegimeObj,function-method}
\alias{.newRegimeObj,list-method}
\title{Create a New \code{RegimeObj} Object}
\usage{
.newRegimeObj(object)

\S4method{.newRegimeObj}{`function`}(object)

\S4method{.newRegimeObj}{list}(object)
}
\arguments{
\item{object}{A function defining the treatment regime}
}
\description{
Calls newRegime and stores object in @regime.
}
\keyword{internal}
