% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_class_RWL.R
\docType{methods}
\name{.newRWL,Kernel-method}
\alias{.newRWL,Kernel-method}
\title{Complete a Residual Weighted Learning Analysis}
\usage{
\S4method{.newRWL}{Kernel}(moPropen, moMain, responseType, data, response,
  txName, lambdas, cvFolds, surrogate, guess, kernel, fSet, suppress, ...)
}
\arguments{
\item{moPropen}{modelObj for propensity modeling}

\item{moMain}{modelObj for main effects}

\item{responseType}{Character indicating type of response}

\item{data}{data.frame of covariates}

\item{response}{vector of responses}

\item{txName}{treatment variable column header in data}

\item{lambdas}{tuning parameter(s)}

\item{cvFolds}{number of cross-validation folds}

\item{surrogate}{Surrogate object}

\item{guess}{optional numeric vector providing starting values for
optimization methods}

\item{kernel}{Kernel object}

\item{fSet}{Function or NULL defining subsets}

\item{suppress}{T/F indicating if prints to screen are executed}

\item{...}{Additional inputs for optimization}

\item{txVec}{treatment vector recast as +/- 1}
}
\value{
An RWL object
}
\description{
Complete a Residual Weighted Learning Analysis
}
\keyword{internal}
