% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_class_RWL.R
\name{newRWL}
\alias{newRWL}
\alias{.newRWL,SubsetList-method}
\title{Complete a Residual Weighted Learning Analysis}
\usage{
.newRWL(kernel, ...)

\S4method{.newRWL}{SubsetList}(
  moPropen,
  moMain,
  responseType,
  data,
  response,
  txName,
  lambdas,
  cvFolds,
  surrogate,
  guess,
  kernel,
  fSet,
  suppress,
  ...
)
}
\arguments{
\item{kernel}{A Kernel object}
}
\description{
Complete a Residual Weighted Learning Analysis
}
\keyword{internal}
