% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxObj.R, R/B_TxInfoNoSubsets.R,
%   R/B_TxInfoWithSubsets.R, R/B_TxInfoList.R
\name{newTxObj}
\alias{newTxObj}
\alias{.newTxObj,ANY,character-method}
\alias{.newTxObj,NULL,character-method}
\alias{.newTxObj,function,character-method}
\alias{.newTxObj,list,list-method}
\alias{.newTxObj,NULL,list-method}
\title{Create \code{TxObj} Object}
\usage{
.newTxObj(fSet, txName, ...)

\S4method{.newTxObj}{ANY,character}(fSet, txName, ...)

\S4method{.newTxObj}{`NULL`,character}(fSet, txName, data, suppress, verify = TRUE)

\S4method{.newTxObj}{`function`,character}(fSet, txName, data, suppress, verify = TRUE)

\S4method{.newTxObj}{list,list}(fSet, txName, data, suppress, verify = TRUE)

\S4method{.newTxObj}{`NULL`,list}(fSet, txName, data, suppress, verify = TRUE)
}
\description{
Creates appropriate \code{TxObj} based on class of fSet and txName.
}
\keyword{internal}
