% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxSubset.R, R/B_TxSubsetInteger.R,
%   R/B_TxSubsetFactor.R
\docType{methods}
\name{newTxSubset}
\alias{newTxSubset}
\alias{.newTxSubset,function,ANY-method}
\alias{.newTxSubset,function,integer-method}
\alias{.newTxSubset,function,character-method}
\title{Create \code{TxSubset} Object}
\usage{
.newTxSubset(fSet, superset, ...)

\S4method{.newTxSubset}{`function`,ANY}(fSet, data, superset, verify,
  suppress, ...)

\S4method{.newTxSubset}{`function`,integer}(fSet, superset, txName, data,
  verify, ...)

\S4method{.newTxSubset}{`function`,character}(fSet, superset, txName, data,
  verify, ...)
}
\description{
Processes input to determine ptsSubset and singleton to create a
  \code{TxSubset} object
}
\keyword{internal}
