% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_TypedFit.R, R/C_TypedFit_SubsetList.R,
%   R/C_TypedFit_fSet.R
\name{newTypedFit}
\alias{newTypedFit}
\alias{.newTypedFit,modelObj,TxInfoNoSubsets-method}
\alias{.newTypedFit,ModelObj_SubsetList,TxInfoWithSubsets-method}
\alias{.newTypedFit,modelObj,TxInfoWithSubsets-method}
\title{Complete a Regression Step}
\usage{
.newTypedFit(modelObj, txObj, ...)

\S4method{.newTypedFit}{modelObj,TxInfoNoSubsets}(modelObj, txObj, response, data, type, suppress)

\S4method{.newTypedFit}{ModelObj_SubsetList,TxInfoWithSubsets}(modelObj, txObj, data, response, type, suppress)

\S4method{.newTypedFit}{modelObj,TxInfoWithSubsets}(modelObj, txObj, data, response, type, suppress)
}
\arguments{
\item{modelObj}{A modeling object}

\item{txObj}{A TxObj object}

\item{...}{Any optional additional input.}
}
\description{
This function completes a regression step and stores a character object
  used to identify the purpose of the step, such as a propensity or
  outcome regression.
}
\keyword{internal}
