% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_TypedFitObj.R
\name{newTypedFitObj}
\alias{newTypedFitObj}
\alias{.newTypedFitObj,ANY-method}
\title{Create a new \code{TypedFitObj} object}
\usage{
.newTypedFitObj(modelObj, txObj, ...)

\S4method{.newTypedFitObj}{ANY}(modelObj, txObj, response, data, type, suppress)
}
\arguments{
\item{modelObj}{A modeling object}

\item{txObj}{A TxObj object}

\item{...}{Any optional additional input.}
}
\description{
Call newTypedFit and stores result in @fit
}
\keyword{internal}
