% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_class_OptimalClass.R
\docType{methods}
\name{OptimalClass-methods}
\alias{OptimalClass-methods}
\alias{print,OptimalClass-method}
\alias{show,OptimalClass-method}
\title{Methods Available for Objects of Class \code{OptimalClass}}
\usage{
\S4method{print}{OptimalClass}(x, ...)

\S4method{show}{OptimalClass}(object)
}
\description{
Methods Available for Objects of Class \code{OptimalClass}
}
\keyword{internal}
