% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_OptimalInfo.R
\docType{class}
\name{OptimalInfo-class}
\alias{OptimalInfo-class}
\title{Class \code{OptimalInfo}}
\description{
Class \code{OptimalInfo} stores the estimated optimal tx, decision functions,
and estimated value.
}
\section{Slots}{

\describe{
\item{\code{optimalTx}}{a vector of the estimated optimal tx}

\item{\code{estimatedValue}}{a vector of the estimated value}

\item{\code{decisionFunc}}{a vector or matrix containing the values used to determine
@optimalTx (if applicable)}
}}

