% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_OptimalInfo.R
\docType{methods}
\name{OptimalInfo-methods}
\alias{OptimalInfo-methods}
\alias{estimator,OptimalInfo-method}
\alias{optTx,OptimalInfo,missing-method}
\alias{optTx,OptimalInfo,ANY-method}
\alias{print,OptimalInfo-method}
\alias{show,OptimalInfo-method}
\alias{summary,OptimalInfo-method}
\title{Methods Available for Objects of Class \code{OptimalInfo}}
\usage{
\S4method{estimator}{OptimalInfo}(x)

\S4method{optTx}{OptimalInfo,missing}(x, newdata, ...)

\S4method{optTx}{OptimalInfo,ANY}(x, newdata, ...)

\S4method{print}{OptimalInfo}(x, ...)

\S4method{show}{OptimalInfo}(object)

\S4method{summary}{OptimalInfo}(object, ...)
}
\description{
Methods Available for Objects of Class \code{OptimalInfo}

\code{estimator(x)} 
  defines the estimated value to be the mean of the vector stored in 
  @estimatedValue

\code{optTx(x)} 
  returns the contents of @optimalTx and @decisionFunc as a list

\code{optTx(x, newdata)} 
  returns an error

\code{print(x)}
  Prints a summary table of the recommended tx for the training data and the
  estimated value

\code{show(object)}
  Displays a summary table of the recommended tx for the training data and
  the estimated value

\code{summary(object)}
  Returns a list containing a summary table of the recommended tx for the 
  training data and the estimated value
}
\keyword{internal}
