% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_OptimalObj.R
\docType{class}
\name{OptimalObj-class}
\alias{OptimalObj-class}
\title{Class \code{OptimalObj}}
\description{
Class \code{OptimalObj} stores the estimated optimal Tx, decision functions
and estimated value under a common name.
}
\section{Slots}{

\describe{
\item{\code{optimal}}{ANY - must be \code{OptimalInfo} or 
\code{DecisionPointList} of \code{OptimalInfo}}
}}

\keyword{internal}
