% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_OptimalObj.R
\docType{methods}
\name{OptimalObj-methods}
\alias{OptimalObj-methods}
\alias{estimator,OptimalObj-method}
\alias{optTx,OptimalObj,missing-method}
\alias{optTx,OptimalObj,data.frame-method}
\alias{print,OptimalObj-method}
\alias{show,OptimalObj-method}
\alias{summary,OptimalObj-method}
\title{Methods Available for Objects of Class \code{OptimalObj}}
\usage{
\S4method{estimator}{OptimalObj}(x)

\S4method{optTx}{OptimalObj,missing}(x, newdata, ...)

\S4method{optTx}{OptimalObj,data.frame}(x, newdata, ...)

\S4method{print}{OptimalObj}(x, ...)

\S4method{show}{OptimalObj}(object)

\S4method{summary}{OptimalObj}(object, ...)
}
\description{
Methods Available for Objects of Class \code{OptimalObj}

\code{estimator(x)} 
  retrieves the estimated value obtained by a statistical method.
  Method called determined by class of @optimal.

\code{optTx(x)} 
  returns the estimated decision function and/or optimal tx
  Method called determined by class of @optimal.

\code{optTx(x, newdata)} 
  returns an error

\code{print(x)}
  Prints summary information regarding recommended tx and the estimated
  value. Method called determined by class of @optimal.

\code{show(object)}
  Displays summary information regarding recommended tx and the estimated
  value. Method called determined by class of @optimal.

\code{summary(object)}
  Returns a summary of estimated decision functions and/or optimal tx. 
  Method called determined by class of @optimal.
}
\keyword{internal}
