\name{OptimalSeq-class}
\docType{class}
\alias{OptimalSeq-class}

\title{Class \code{"OptimalSeq"}}
\description{
VIRTUAL class for value search method with IPWE or AIPWE.
}
\section{Objects from the Class}{
A virtual Class: No object may be created from it.
}
\section{Extends}{
Class \code{"\link{DynTxRegime-class}"} directly.
}
\section{Methods}{
  No methods identified with signature \code{"OptimalSeq"}.
}
\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\keyword{internal}
