% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_class_OptimalSeq.R
\name{OptimalSeq-methods}
\alias{OptimalSeq-methods}
\alias{print,OptimalSeq-method}
\alias{show,OptimalSeq-method}
\title{Methods Available for Objects of Class \code{OptimalSeq}}
\usage{
\S4method{print}{OptimalSeq}(x, ...)

\S4method{show}{OptimalSeq}(object)
}
\description{
Methods Available for Objects of Class \code{OptimalSeq}

\code{print(x)}
  prints main results of a coarsened/missing data analysis

\code{show(object)}
  displays main results of a coarsened/missing data analysis
}
\keyword{internal}
