% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_class_OptimalSeqCoarsened.R
\docType{class}
\name{OptimalSeqCoarsened-class}
\alias{OptimalSeqCoarsened-class}
\title{Class Contains Results for the Coarsened Data IPW/AIPW Method}
\description{
Methods for multiple decision point analyses. Class inherits directly from
 \code{OptimalSeq} and all methods defined for objects of class \code{OptimaSeq}
 are defined for this class.
}
