% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_class_OptimalSeqCoarsened.R
\docType{methods}
\name{OptimalSeqCoarsened-methods}
\alias{OptimalSeqCoarsened-methods}
\alias{Call,OptimalSeqCoarsened-method}
\alias{coef,OptimalSeqCoarsened-method}
\alias{DTRstep,OptimalSeqCoarsened-method}
\alias{estimator,OptimalSeqCoarsened-method}
\alias{fitObject,OptimalSeqCoarsened-method}
\alias{genetic,OptimalSeqCoarsened-method}
\alias{optTx,OptimalSeqCoarsened,data.frame-method}
\alias{optTx,OptimalSeqCoarsened,missing-method}
\alias{outcome,OptimalSeqCoarsened-method}
\alias{plot,OptimalSeqCoarsened,ANY-method}
\alias{print,OptimalSeqCoarsened-method}
\alias{propen,OptimalSeqCoarsened-method}
\alias{regimeCoef,OptimalSeqCoarsened-method}
\alias{show,OptimalSeqCoarsened-method}
\alias{summary,OptimalSeqCoarsened-method}
\title{Methods Available for Objects of Class \code{OptimalSeqCoarsened}}
\usage{
\S4method{Call}{OptimalSeqCoarsened}(name, ...)

\S4method{coef}{OptimalSeqCoarsened}(object, ...)

\S4method{DTRstep}{OptimalSeqCoarsened}(object)

\S4method{estimator}{OptimalSeqCoarsened}(x, ...)

\S4method{fitObject}{OptimalSeqCoarsened}(object, ...)

\S4method{genetic}{OptimalSeqCoarsened}(object, ...)

\S4method{optTx}{OptimalSeqCoarsened,data.frame}(x, newdata, ..., dp = 1)

\S4method{optTx}{OptimalSeqCoarsened,missing}(x, newdata, ...)

\S4method{outcome}{OptimalSeqCoarsened}(object, ...)

\S4method{plot}{OptimalSeqCoarsened,ANY}(x, suppress = FALSE, ...)

\S4method{print}{OptimalSeqCoarsened}(x, ...)

\S4method{propen}{OptimalSeqCoarsened}(object, ...)

\S4method{regimeCoef}{OptimalSeqCoarsened}(object, ...)

\S4method{show}{OptimalSeqCoarsened}(object)

\S4method{summary}{OptimalSeqCoarsened}(object, ...)
}
\description{
Methods Available for Objects of Class \code{OptimalSeqCoarsened}

\code{Call(name)}
  returns the unevaluated call to method

\code{coef(object)}
  retrieves coefficients of model functions. Calls method defined for 
  \code{OptimalSeq}.

\code{DTRstep(x)}
  print statement indicating the coarsened data perspective

\code{estimator(x)}
  retrieves the estimated value. Calls method defined for 
  \code{OptimalSeq}.

\code{fitObject(object)}
  retrieves value objects of model functions. Calls method defined for 
  \code{OptimalSeq}.

\code{genetic(object)}
  retrieves genetic algorithm results. Calls method defined for 
  \code{OptimalSeq}.

\code{optTx(x,newdata)}
  estimates optimal tx. Calls method defined for \code{OptimalSeq}.

\code{optTx(x)}
  retrieves the optimal tx. Calls method defined for \code{OptimalSeq}.

\code{outcome(object)}
  retrieves value object returned by outcome model functions. Calls method 
  defined for \code{OptimalSeq}.

\code{plot(x,suppress)}
  generates plot for model functions. Calls method defined for 
  \code{OptimalSeq}.

\code{print(x)}
  Extends method defined for \code{OptimalSeq} to include DTRStep()

\code{propen(object)}
  retrieves value object returned by propensity model functions. Calls method 
  defined for \code{OptimalSeq}.

\code{regimeCoef(object)}
  retrieves estimated tx regime parameters. Calls method defined for 
  \code{OptimalSeq}.

\code{show(object)}
  Extends method defined for \code{OptimalSeq} to include DTRStep()

\code{summary(object)}
  retrieves summary information. Calls method defined for \code{OptimalSeq}.
}
\keyword{internal}
