% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_class_OptimalSeqMissing.R
\docType{class}
\name{OptimalSeqMissing-class}
\alias{OptimalSeqMissing-class}
\title{Class Contains Results for the Missing Data IPW/AIPW Method}
\description{
Methods for single decision point analyses. Class inherits directly from
\code{OptimalSeq} and all methods defined for objects of class \code{OptimaSeq}
are defined for this class.
}
