% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_OptimBasic.R
\docType{methods}
\name{OptimBasic-methods}
\alias{OptimBasic-methods}
\alias{optimObj,OptimBasic-method}
\alias{.predictOptimalTx,OptimBasic,matrix-method}
\alias{.predictOptimalTx,OptimBasic,data.frame-method}
\alias{.predictOptimalTx,OptimBasic,missing-method}
\alias{print,OptimBasic-method}
\alias{regimeCoef,OptimBasic-method}
\alias{show,OptimBasic-method}
\alias{summary,OptimBasic-method}
\title{Methods Available for Objects of Class \code{OptimBasic}}
\usage{
\S4method{optimObj}{OptimBasic}(object, ...)

\S4method{.predictOptimalTx}{OptimBasic,matrix}(x, newdata)

\S4method{.predictOptimalTx}{OptimBasic,data.frame}(x, newdata, ...)

\S4method{.predictOptimalTx}{OptimBasic,missing}(x, newdata, ...)

\S4method{print}{OptimBasic}(x, ...)

\S4method{regimeCoef}{OptimBasic}(object, ...)

\S4method{show}{OptimBasic}(object)

\S4method{summary}{OptimBasic}(object, ...)
}
\description{
Methods Available for Objects of Class \code{OptimBasic}
}
\keyword{internal}
