% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_OptimKernel.R
\name{OptimKernel-methods}
\alias{OptimKernel-methods}
\alias{.predictOptimalTx,OptimKernel,matrix-method}
\alias{.predictOptimalTx,OptimKernel,data.frame-method}
\alias{.predictOptimalTx,OptimKernel,missing-method}
\title{Methods Available for Objects of Class \code{OptimKernel}}
\usage{
\S4method{.predictOptimalTx}{OptimKernel,matrix}(x, newdata)

\S4method{.predictOptimalTx}{OptimKernel,data.frame}(x, newdata, ...)

\S4method{.predictOptimalTx}{OptimKernel,missing}(x, newdata, ...)
}
\description{
Methods Available for Objects of Class \code{OptimKernel}
}
\keyword{internal}
