% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_OptimObj.R
\docType{class}
\name{OptimObj-class}
\alias{OptimObj-class}
\title{Class \code{OptimObj}}
\description{
Class \code{OptimObj} stores the optimization results under a common name
  for weighted learning methods.
}
\section{Slots}{

\describe{
\item{\code{optim}}{ANY - expected to be \code{OptimBasic} or \code{OptimKernel}}
}}

\keyword{internal}
