% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_OptimObj.R
\name{OptimObj-methods}
\alias{OptimObj-methods}
\alias{optimObj,OptimObj-method}
\alias{.predictOptimalTx,OptimObj,matrix-method}
\alias{.predictOptimalTx,OptimObj,data.frame-method}
\alias{.predictOptimalTx,OptimObj,missing-method}
\alias{print,OptimObj-method}
\alias{regimeCoef,OptimObj-method}
\alias{show,OptimObj-method}
\alias{summary,OptimObj-method}
\title{Methods Available for Objects of Class \code{OptimObj}}
\usage{
\S4method{optimObj}{OptimObj}(object, ...)

\S4method{.predictOptimalTx}{OptimObj,matrix}(x, newdata, ...)

\S4method{.predictOptimalTx}{OptimObj,data.frame}(x, newdata, ...)

\S4method{.predictOptimalTx}{OptimObj,missing}(x, newdata, ...)

\S4method{print}{OptimObj}(x, ...)

\S4method{regimeCoef}{OptimObj}(object, ...)

\S4method{show}{OptimObj}(object)

\S4method{summary}{OptimObj}(object, ...)
}
\description{
Methods Available for Objects of Class \code{OptimObj}
}
\keyword{internal}
