% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_OptimBasic.R
\name{optimObj}
\alias{optimObj}
\title{Extract Optimization Results}
\usage{
optimObj(object, ...)

## S4 method for signature 'OWL'
optimObj(object, ...)

## S4 method for signature 'RWL'
optimObj(object, ...)

## S4 method for signature 'BOWL'
optimObj(object, ...)

## S4 method for signature 'EARL'
optimObj(object, ...)
}
\arguments{
\item{object}{A value object returned by a statistical method of DynTxRegime
that uses optimization to estimate regime parameters.}

\item{...}{Ignored.}
}
\description{
Retrieves the value object returned by the optimization method for
  weighted learning methods.
}
