% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_OptimStep.R
\docType{class}
\name{OptimStep-class}
\alias{OptimStep-class}
\title{Class \code{OptimStep}
Class \code{OptimStep} holds results of a combined cross-validation and final
  optimization step for weighted learning methods.}
\description{
Class \code{OptimStep}
Class \code{OptimStep} holds results of a combined cross-validation and final
  optimization step for weighted learning methods.
}
\keyword{internal}
