% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_OptimStep.R
\name{OptimStep-methods}
\alias{OptimStep-methods}
\alias{Call,OptimStep-method}
\alias{cvInfo,OptimStep-method}
\alias{estimator,OptimStep-method}
\alias{optimObj,OptimStep-method}
\alias{optTx,OptimStep,missing-method}
\alias{.predictOptimalTx,OptimStep,matrix-method}
\alias{.predictOptimalTx,OptimStep,data.frame-method}
\alias{.predictOptimalTx,OptimStep,missing-method}
\alias{print,OptimStep-method}
\alias{regimeCoef,OptimStep-method}
\alias{show,OptimStep-method}
\alias{summary,OptimStep-method}
\title{Methods Available for Objects of Class \code{OptimStep}}
\usage{
\S4method{Call}{OptimStep}(name, ...)

\S4method{cvInfo}{OptimStep}(object)

\S4method{estimator}{OptimStep}(x, ...)

\S4method{optimObj}{OptimStep}(object)

\S4method{optTx}{OptimStep,missing}(x, newdata, ...)

\S4method{.predictOptimalTx}{OptimStep,matrix}(x, newdata, ...)

\S4method{.predictOptimalTx}{OptimStep,data.frame}(x, newdata, ...)

\S4method{.predictOptimalTx}{OptimStep,missing}(x, newdata, ...)

\S4method{print}{OptimStep}(x, ...)

\S4method{regimeCoef}{OptimStep}(object)

\S4method{show}{OptimStep}(object)

\S4method{summary}{OptimStep}(object, ...)
}
\description{
Methods Available for Objects of Class \code{OptimStep}
}
\keyword{internal}
