% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_OptimStep.R
\name{OptimStep}
\alias{OptimStep}
\alias{.OptimStep}
\title{Complete Cross-Validation Step and Final Optimization}
\usage{
.OptimStep(methodObject, lambdas, cvFolds, txVec, suppress, ...)
}
\arguments{
\item{methodObject}{Object parameters for weighted learning method}

\item{lambdas}{tuning parameter}

\item{cvFolds}{number of cross-validation folds}

\item{suppress}{integer indicating screen printing preferences}
}
\description{
Complete Cross-Validation Step and Final Optimization
}
\keyword{internal}
