% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeNoFit.R
\name{outcome}
\alias{outcome}
\title{Retrieve Outcome Regression Analysis}
\usage{
outcome(object, ...)
}
\arguments{
\item{object}{A value object returned by a statistical method of DynTxRegime.}

\item{...}{Ignored.}
}
\description{
For statistical methods that require an outcome regression analysis,
  the value object returned by the modeling function(s) is retrieved.
}
\details{
Methods are defined for all statistical methods implemented in DynTxRegime
  that use outcome regression.
}
