% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeIterateFit.R
\docType{class}
\name{OutcomeIterateFit-class}
\alias{OutcomeIterateFit-class}
\title{Class \code{OutcomeIterateFit}}
\description{
Class \code{OutcomeIterateFit} is a an outcome regression step completed
  using the iterative algorithm.
}
\section{Slots}{

\describe{
\item{\code{fitObjC}}{Contrast Result}

\item{\code{fitObjM}}{Main Effects Result}
}}

\keyword{internal}
