% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeIterateFit.R
\docType{methods}
\name{OutcomeIterateFit-methods}
\alias{OutcomeIterateFit-methods}
\alias{coef,OutcomeIterateFit-method}
\alias{fitObject,OutcomeIterateFit-method}
\alias{outcome,OutcomeIterateFit-method}
\alias{plot,OutcomeIterateFit,ANY-method}
\alias{predict,OutcomeIterateFit-method}
\alias{.predictAll,OutcomeIterateFit,data.frame-method}
\alias{print,OutcomeIterateFit-method}
\alias{show,OutcomeIterateFit-method}
\alias{summary,OutcomeIterateFit-method}
\title{Methods Available for Objects of Class \code{OutcomeIterateFit}}
\usage{
\S4method{coef}{OutcomeIterateFit}(object, ...)

\S4method{fitObject}{OutcomeIterateFit}(object, ...)

\S4method{outcome}{OutcomeIterateFit}(object, ...)

\S4method{plot}{OutcomeIterateFit,ANY}(x, suppress = FALSE, ...)

\S4method{predict}{OutcomeIterateFit}(object, ...)

\S4method{.predictAll}{OutcomeIterateFit,data.frame}(object, newdata, ...)

\S4method{print}{OutcomeIterateFit}(x, ...)

\S4method{show}{OutcomeIterateFit}(object)

\S4method{summary}{OutcomeIterateFit}(object, ...)
}
\description{
Methods call equivalently named methods defined for \code{OutcomeSimpleFit},
  \code{OutcomeSimpleFit_fSet}, or \code{OutcomeSimpleFit_SubsetList}.
Exact method dispatched depends on classes of @fitObjC and @fitObjM.
When a value object is returned, it is a list.

\code{.predictAll(object, newdata)} combines the two components into a
  single optimal tx and decision function
}
\keyword{internal}
