% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeNoFit.R
\docType{class}
\name{OutcomeNoFit-class}
\alias{OutcomeNoFit-class}
\title{Class \code{OutcomeNoFit}}
\description{
Class \code{OutcomeNoFit} designates that an outcome regression step
  was not performed. This acts as a place holder for IPW based methods.
}
\keyword{internal}
