% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeObj.R
\docType{class}
\name{OutcomeObj-class}
\alias{OutcomeObj-class}
\title{Class \code{OutcomeObj}}
\description{
Class \code{OutcomeObj} groups outcome regression results under a common
 name
}
\section{Slots}{

\describe{
\item{\code{outcome}}{ANY - expected to be \code{OutcomeNoFit}, 
\code{OutcomeSimpleFit},
\code{OutcomeSimpleFit_fSet}, \code{OutcomeSimpleFit_SubsetList}, 
\code{OutcomeIterateFit}, or \code{DecisionPointList}.}
}}

\keyword{internal}
