% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeObj.R
\name{OutcomeObj-methods}
\alias{OutcomeObj-methods}
\alias{coef,OutcomeObj-method}
\alias{fitObject,OutcomeObj-method}
\alias{outcome,OutcomeObj-method}
\alias{plot,OutcomeObj,ANY-method}
\alias{.predictAll,OutcomeObj,data.frame-method}
\alias{.predictMu,OutcomeObj,data.frame-method}
\alias{predict,OutcomeObj-method}
\alias{print,OutcomeObj-method}
\alias{show,OutcomeObj-method}
\alias{summary,OutcomeObj-method}
\title{Methods Available for Objects of Class \code{OutcomeObj}}
\usage{
\S4method{coef}{OutcomeObj}(object, ...)

\S4method{fitObject}{OutcomeObj}(object, ...)

\S4method{outcome}{OutcomeObj}(object, ...)

\S4method{plot}{OutcomeObj,ANY}(x, suppress = FALSE, ...)

\S4method{.predictAll}{OutcomeObj,data.frame}(object, newdata, ...)

\S4method{.predictMu}{OutcomeObj,data.frame}(object, data, ...)

\S4method{predict}{OutcomeObj}(object, ...)

\S4method{print}{OutcomeObj}(x, ...)

\S4method{show}{OutcomeObj}(object)

\S4method{summary}{OutcomeObj}(object, ...)
}
\description{
Most value objects returned are a list with one element 'outcome'.
Methods dispatched and objects returned in the element 'outcome'
  depend on class of @outcome.
Exceptions are noted below.

\code{outcome(object)} does not return the overarching list
  structure, but only the contents of list[[ outcome ]].

\code{plot(x)} concatenated 'outcome' to the title if suppress = FALSE.

\code{.predictAll(object, newdata)} does not return the overarching list
  structure, but only the contents of list[[ outcome ]].

\code{.predictMu(object, newdata)}
  predicts outcome for all tx options.
  Returns the matrix of outcomes predicted for all tx. 
  Predicted outcomes for tx not available to a pt are NA.

\code{predict(object)} does not return the overarching list
  structure, but only the contents of list[[ outcome ]].
}
\keyword{internal}
