\name{OutcomeRegression-class}
\Rdversion{1.1}
\docType{class}
\alias{OutcomeRegression-class}

\title{Class \code{"OutcomeRegression"}}

\description{
Class to identify an object as an outcome regression.
}

\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}

\section{Methods}{
No methods defined with class \code{"OutcomeRegression"} in the signature.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("OutcomeRegression")
}

\keyword{internal}
