% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeSimpleFit.R
\docType{class}
\name{OutcomeSimpleFit-class}
\alias{OutcomeSimpleFit-class}
\title{Class \code{OutcomeSimpleFit}}
\description{
Class \code{OutcomeSimpleFit} is a \code{TypedFit} identified as being
  for an outcome regression step.
}
\keyword{internal}
