% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeSimpleFit_fSet.R
\docType{class}
\name{OutcomeSimpleFit_fSet-class}
\alias{OutcomeSimpleFit_fSet-class}
\title{Class \code{OutcomeSimpleFit_fSet}}
\description{
Class \code{OutcomeSimpleFit_fSet} is a \code{TypedFit_fSet} identified as
  being for an outcome regression step.
}
\keyword{internal}
