% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeSimpleFit_fSet.R
\docType{methods}
\name{OutcomeSimpleFit_fSet-methods}
\alias{OutcomeSimpleFit_fSet-methods}
\alias{outcome,OutcomeSimpleFit_fSet-method}
\alias{.predictAll,OutcomeSimpleFit_fSet,data.frame-method}
\title{Methods Available for Objects of Class \code{OutcomeSimpleFit_fSet}}
\usage{
\S4method{outcome}{OutcomeSimpleFit_fSet}(object, ...)

\S4method{.predictAll}{OutcomeSimpleFit_fSet,data.frame}(object, newdata,
  ...)
}
\description{
Methods call equivalently named methods defined for \code{TypedFit_fSet}

\code{.predictAll(object, newdata)}
  predicts outcome for all tx options.
  Returns a list containing 'optimalTx' the tx yielding the largest 
  predicted outcome and 'decisionFunc' the matrix of outcomes predicted
  for all tx. 
  Predicted outcomes for tx not available to a pt are NA.
}
\keyword{internal}
