% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeSimpleFit_SubsetList.R
\docType{class}
\name{OutcomeSimpleFit_SubsetList-class}
\alias{OutcomeSimpleFit_SubsetList-class}
\title{Class \code{OutcomeSimpleFit_SubsetList}}
\description{
Class \code{OutcomeSimpleFit_SubsetList} is a \code{TypedFit_SubsetList} 
  identified as being for an outcome regression step.
}
\keyword{internal}
