% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeSimpleFit_SubsetList.R
\docType{methods}
\name{OutcomeSimpleFit_SubsetList-methods}
\alias{OutcomeSimpleFit_SubsetList-methods}
\alias{outcome,OutcomeSimpleFit_SubsetList-method}
\alias{predict,OutcomeSimpleFit_SubsetList-method}
\alias{.predictAll,OutcomeSimpleFit_SubsetList,data.frame-method}
\title{Methods Available for Objects of Class \code{OutcomeSimpleFit_SubsetList}}
\usage{
\S4method{outcome}{OutcomeSimpleFit_SubsetList}(object, ...)

\S4method{predict}{OutcomeSimpleFit_SubsetList}(object, newdata, ...)

\S4method{.predictAll}{OutcomeSimpleFit_SubsetList,data.frame}(object,
  newdata, ...)
}
\description{
Methods call equivalently named methods defined for \code{TypedFit_SubsetList}

\code{.predictAll(object, newdata)}
  predicts outcome for all tx options.
  Returns a list containing 'optimalTx' the tx yielding the largest 
  predicted outcome and 'decisionFunc' the matrix of outcomes predicted
  for all tx. 
  Predicted outcomes for tx not available to a pt are NA.
}
\keyword{internal}
