% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_class_OWL.R
\name{OWL-methods}
\alias{OWL-methods}
\alias{print,OWL-method}
\alias{show,OWL-method}
\title{Methods Available for Objects of Class \code{OWL}}
\usage{
\S4method{print}{OWL}(x, ...)

\S4method{show}{OWL}(object)
}
\description{
Methods Available for Objects of Class \code{OWL}
}
\keyword{internal}
