% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_generics.R
\name{plot}
\alias{plot}
\title{Generates Plots as Defined by Modeling Functions}
\arguments{
\item{x}{Value object returned by a statistical method}

\item{y}{Ignored}

\item{suppress}{T/F indicating if titles should be concatenated with
information indicating the specific regression step}

\item{...}{Optional additional inputs}
}
\description{
Calls plot() method for all regression steps of a statistical method
}
\details{
Methods are defined for all statistical methods implemented in DynTxRegime.
}
