% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_PolyKernel.R
\docType{class}
\name{PolyKernel-class}
\alias{PolyKernel-class}
\title{Class \code{PolyKernel}}
\description{
Class \code{PolyKernel} holds information regarding decision function 
  when kernel is polynomial
}
\keyword{internal}
