% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_PolyKernel.R
\name{PolyKernel-methods}
\alias{PolyKernel-methods}
\alias{.kernel,PolyKernel,matrix,matrix-method}
\alias{print,PolyKernel-method}
\alias{show,PolyKernel-method}
\alias{summary,PolyKernel-method}
\title{Methods Available for Objects of Class \code{PolyKernel}}
\usage{
\S4method{.kernel}{PolyKernel,matrix,matrix}(object, x1, x2, ...)

\S4method{print}{PolyKernel}(x, ...)

\S4method{show}{PolyKernel}(object)

\S4method{summary}{PolyKernel}(object, ...)
}
\description{
Methods Available for Objects of Class \code{PolyKernel}
}
\keyword{internal}
