% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityFit.R
\name{propen}
\alias{propen}
\title{Retrieve Propensity Regression Analysis}
\usage{
propen(object, ...)
}
\arguments{
\item{object}{A value object returned by a statistical method of DynTxRegime.}

\item{...}{Ignored.}
}
\description{
For statistical methods that require a propensity regression analysis,
  the value object returned by the modeling function(s) is retrieved.
}
\details{
Methods are defined for all statistical methods implemented in DynTxRegime
  that use propensity regression.
}
