% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityFit.R
\name{PropensityFit-methods}
\alias{PropensityFit-methods}
\alias{coef,PropensityFit-method}
\alias{fitObject,PropensityFit-method}
\alias{plot,PropensityFit,ANY-method}
\alias{predict,PropensityFit-method}
\alias{.predictAll,PropensityFit,data.frame-method}
\alias{print,PropensityFit-method}
\alias{propen,PropensityFit-method}
\alias{show,PropensityFit-method}
\alias{summary,PropensityFit-method}
\title{Methods Available for Objects of Class \code{PropensityFit}}
\usage{
\S4method{coef}{PropensityFit}(object, ...)

\S4method{fitObject}{PropensityFit}(object, ...)

\S4method{plot}{PropensityFit,ANY}(x, suppress = FALSE, ...)

\S4method{predict}{PropensityFit}(object, ...)

\S4method{.predictAll}{PropensityFit,data.frame}(object, newdata, suppress = TRUE)

\S4method{print}{PropensityFit}(x, ...)

\S4method{propen}{PropensityFit}(object, ...)

\S4method{show}{PropensityFit}(object)

\S4method{summary}{PropensityFit}(object, ...)
}
\description{
Methods call equivalently named methods defined for \code{TypedFit}

\code{.predictAll(object, newdata)}
  predicts propensity for all tx options.
  Returns a matrix of propensities predicted for all tx.
}
\keyword{internal}
