\name{PropensityFit_DecisionPointList-class}
\docType{class}
\alias{PropensityFit_DecisionPointList-class}

\title{Class \code{"PropensityFit_DecisionPointList"}}
\description{
Propensity score regression for multiple-decision-point
analyses.
}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("PropensityFit_DecisionPointList", ...)}.
These objects are for convenience in package development
and should not be created by users.
}

\section{Extends}{
Classes \code{"\link{DecisionPointList-class}"}, \code{"\link{PropensityRegression-class}"},
\code{"\link{SubsetsNotModeled-class}"}, directly.
Classes \code{"\link{MultipleDecisionPoint-class}"} and \code{"\link{List-class}"} by class \code{"\link{DecisionPointList-class}"}, distance 2.
}

\section{Methods}{
No methods identified with \code{"PropensityFit_DecisionPointList"} in signature.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("PropensityFit_DecisionPointList")
}

\keyword{internal}
