% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityFit_fSet.R
\docType{class}
\name{PropensityFit_fSet-class}
\alias{PropensityFit_fSet-class}
\title{Class \code{PropensityFit_fSet}}
\description{
Class \code{PropensityFit_fSet} is a \code{TypedFit_fSet} identified as being
  for a propensity regression step.
}
\section{Slots}{

\describe{
\item{\code{small}}{A logical TRUE indicates that the smallest valued tx is
missing; FALSE indicates that the largest valued tx is missing}

\item{\code{levs}}{A vector; the set of treatment options included in fit.}
}}

\keyword{internal}
