% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityFit_fSet.R
\docType{methods}
\name{PropensityFit_fSet-methods}
\alias{PropensityFit_fSet-methods}
\alias{coef,PropensityFit_fSet-method}
\alias{fitObject,PropensityFit_fSet-method}
\alias{.predictAll,PropensityFit_fSet,data.frame-method}
\alias{propen,PropensityFit_fSet-method}
\alias{summary,PropensityFit_fSet-method}
\title{Methods Available for Objects of Class \code{PropensityFit_fSet}}
\usage{
\S4method{coef}{PropensityFit_fSet}(object, ...)

\S4method{fitObject}{PropensityFit_fSet}(object, ...)

\S4method{.predictAll}{PropensityFit_fSet,data.frame}(object, newdata,
  suppress = TRUE)

\S4method{propen}{PropensityFit_fSet}(object, ...)

\S4method{summary}{PropensityFit_fSet}(object, ...)
}
\description{
Methods call equivalently named methods defined for \code{TypedFit_fSet}

\code{.predictAll(object, newdata)}
  predicts propensity for all tx options.
  Returns a matrix of propensities predicted for all tx. 
  Tx options not available to a pt are coded as NA.
}
\keyword{internal}
