% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityFit_SubsetList.R
\docType{class}
\name{PropensityFit_SubsetList-class}
\alias{PropensityFit_SubsetList-class}
\title{Class \code{PropensityFit_SubsetList}}
\description{
Class \code{PropensityFit_SubsetList} is a \code{TypedFit_SubsetList} 
  identified as being for a propensity regression step.
}
\section{Slots}{

\describe{
\item{\code{small}}{A logical vector TRUE indicates that the smallest valued tx is
missing; FALSE indicates that the largest valued tx is missing}

\item{\code{levs}}{A list; the set of treatment options included in each fit.}
}}

\keyword{internal}
