% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityFit_SubsetList.R
\name{PropensityFit_SubsetList-methods}
\alias{PropensityFit_SubsetList-methods}
\alias{coef,PropensityFit_SubsetList-method}
\alias{fitObject,PropensityFit_SubsetList-method}
\alias{.predictAll,PropensityFit_SubsetList,data.frame-method}
\alias{propen,PropensityFit_SubsetList-method}
\alias{summary,PropensityFit_SubsetList-method}
\title{Methods Available for Objects of Class \code{PropensityFit_SubsetList}}
\usage{
\S4method{coef}{PropensityFit_SubsetList}(object, ...)

\S4method{fitObject}{PropensityFit_SubsetList}(object, ...)

\S4method{.predictAll}{PropensityFit_SubsetList,data.frame}(object, newdata, suppress = TRUE)

\S4method{propen}{PropensityFit_SubsetList}(object, ...)

\S4method{summary}{PropensityFit_SubsetList}(object, ...)
}
\description{
Most methods call equivalently named methods defined for 
  \code{TypedFit_SubsetList}
}
\keyword{internal}
