\name{PropensityFit_SubsetList_DecisionPointList-class}
\docType{class}
\alias{PropensityFit_SubsetList_DecisionPointList-class}

\title{Class \code{"PropensityFit_SubsetList_DecisionPointList"}}
\description{
Propensity score regression when subset modeling is used
for multiple-decision-point analyses.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("PropensityFit_SubsetList_DecisionPointList", ...)}.
These objects are for convenience in package development
and should not be created by users.
}
\section{Methods}{
No methods identified with \code{"PropensityFit_SubsetList_DecisionPointList"} in signature.
}

\section{Extends}{
Classes 
\code{"\link{DecisionPointList-class},"}
\code{"\link{PropensityRegression-class},"} and
\code{"\link{SubsetsModeled-class}"} directly.

Classes 
\code{"\link{MultipleDecisionPoint-class}"} and 
\code{"\link{List-class}"} by class 
\code{"\link{DecisionPointList-class},"} distance 2.
}

\author{ Shannon T. Holloway <sthollow@ncsu.edu> }

\examples{
showClass("PropensityFit_SubsetList_DecisionPointList")
}

\keyword{internal}
