% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityObj.R
\docType{class}
\name{PropensityObj-class}
\alias{PropensityObj-class}
\title{Class \code{PropensityObj}}
\description{
Class \code{PropensityObj} groups Propensity regression results under a 
  common name.
}
\section{Slots}{

\describe{
\item{\code{Propensity}}{ANY - expected to be \code{PropensitySimpleFit},
\code{PropensitySimpleFit_fSet}, \code{PropensitySimpleFit_SubsetList}, 
or \code{DecisionPointList}.}
}}

\keyword{internal}
