% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityObj.R
\docType{methods}
\name{PropensityObj-methods}
\alias{PropensityObj-methods}
\alias{coef,PropensityObj-method}
\alias{fitObject,PropensityObj-method}
\alias{plot,PropensityObj,ANY-method}
\alias{.predictAll,PropensityObj,data.frame-method}
\alias{predict,PropensityObj-method}
\alias{print,PropensityObj-method}
\alias{propen,PropensityObj-method}
\alias{show,PropensityObj-method}
\alias{summary,PropensityObj-method}
\title{Methods Available for Objects of Class \code{PropensityObj}}
\usage{
\S4method{coef}{PropensityObj}(object, ...)

\S4method{fitObject}{PropensityObj}(object, ...)

\S4method{plot}{PropensityObj,ANY}(x, suppress = FALSE, ...)

\S4method{.predictAll}{PropensityObj,data.frame}(object, newdata, ...)

\S4method{predict}{PropensityObj}(object, ...)

\S4method{print}{PropensityObj}(x, ...)

\S4method{propen}{PropensityObj}(object, ...)

\S4method{show}{PropensityObj}(object)

\S4method{summary}{PropensityObj}(object, ...)
}
\description{
Most value objects returned are a list with one element 'propen'.
Methods dispatched and objects returned in the element 'propen'
  depend on class of @propen.
Exceptions are noted below.

\code{plot(x)} concatenates 'Propensity' to the title if suppress = FALSE.

\code{.predictAll(object, newdata)} does not return the overarching list
  structure, but only the contents of list[[ propen ]].

\code{predict(object)} does not return the overarching list
  structure, but only the contents of list[[ propen ]].

\code{propen(object)} does not return the overarching list
  structure, but only the contents of list[[ propen ]].
}
\keyword{internal}
