\name{PropensityRegression-class}
\Rdversion{1.1}
\docType{class}
\alias{PropensityRegression-class}

\title{Class \code{"PropensityRegression"}}

\description{
Class to identify an object as a propensity regression.
}

\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}

\section{Methods}{
No methods defined with class \code{"PropensityRegression"} in the signature.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("PropensityRegression")
}

\keyword{internal}
