% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_QLearn.R
\docType{class}
\name{QLearnObj-class}
\alias{QLearnObj-class}
\title{Class \code{QLearnObj}}
\description{
Class \code{QLearnObj} contains the results for a Q-Learning step
}
\section{Slots}{

\describe{
\item{\code{outcome}}{The outcome regression analysis}

\item{\code{txInfo}}{The feasible tx information}

\item{\code{optimal}}{The estimated optimal tx, decision function, and value}
}}

\section{Methods For Post-Processing of Regression Analysis}{


 \describe{
   \item{outcome}{: 
      Retrieve value object returned by outcome regression methods.  }
}



 \describe{
   \item{coef}{: 
     Retrieve parameter estimates for all regression steps. }
   \item{fitObject}{: 
      Retrieve value object returned by regression methods. }
   \item{plot}{: 
      Generate plots for regression analyses.  }
}
}

\section{Methods For Accessing Main Results}{


 \describe{
   \item{DTRstep}{: 
      Retrieve description of method used to create object. }
   \item{estimator}{: 
      Retrieve the estimated value of the estimated
      optimal regime for the training data set. }
   \item{optTx}{: 
      Retrieve/predict the estimated decision functions and/or optimal tx. }
   \item{print}{: 
      Print main results of analysis.}
   \item{show}{: 
      Show main results of analysis.}
   \item{summary}{: 
      Retrieve summary information.}
}
}

