% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnFS_VHet.R
\docType{methods}
\name{qqplot}
\alias{qqplot}
\alias{qqplot,IQLearnFS_VHet-method}
\title{Quantile-Quantile Plots}
\usage{
qqplot(x, y, plot.it, xlab, ylab, ...)

  ## S4 method for IQLearnFS_VHet,ANY
  qqplot(x, y, plot.it, xlab, ylab, ...)
}
\arguments{
\item{x}{An object of class IQLearnFS_VHet}

\item{y}{ANY}

\item{plot.it}{logical. Should the result be plotted?}

\item{xlab, ylab}{The 'xlab' and 'ylab' refer to the y and x axes respectively.}

\item{...}{graphical parameters}
}
\description{
QQ plot of the residuals for the interactive Q-Learning algorithm.
}
