% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_RadialKernel.R
\docType{class}
\name{RadialKernel-class}
\alias{RadialKernel-class}
\title{Class \code{RadialKernel}}
\description{
Class \code{RadialKernel} holds information regarding decision function 
  when kernel is radial
}
\keyword{internal}
