% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_RadialKernel.R
\docType{methods}
\name{RadialKernel-methods}
\alias{RadialKernel-methods}
\alias{.kernel,RadialKernel,matrix,matrix-method}
\alias{print,RadialKernel-method}
\alias{show,RadialKernel-method}
\alias{summary,RadialKernel-method}
\title{Methods Available for Objects of Class \code{RadialKernel}}
\usage{
\S4method{.kernel}{RadialKernel,matrix,matrix}(object, x1, x2, ...)

\S4method{print}{RadialKernel}(x, ...)

\S4method{show}{RadialKernel}(object)

\S4method{summary}{RadialKernel}(object, ...)
}
\description{
Methods Available for Objects of Class \code{RadialKernel}
}
\keyword{internal}
