% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_Regime.R
\name{Regime-methods}
\alias{Regime-methods}
\alias{.getNumPars,Regime-method}
\alias{.getParNames,Regime-method}
\alias{.getPars,Regime-method}
\alias{.getRegimeFunction,Regime-method}
\alias{.predictOptimalTx,Regime,data.frame-method}
\alias{print,Regime-method}
\alias{regimeCoef,Regime-method}
\alias{.setPars,Regime,numeric-method}
\alias{show,Regime-method}
\alias{summary,Regime-method}
\title{Methods Available for Objects of Class \code{Regime}}
\usage{
\S4method{.getNumPars}{Regime}(object)

\S4method{.getParNames}{Regime}(object)

\S4method{.getPars}{Regime}(object)

\S4method{.getRegimeFunction}{Regime}(object)

\S4method{.predictOptimalTx}{Regime,data.frame}(x, newdata, ...)

\S4method{print}{Regime}(x, ...)

\S4method{regimeCoef}{Regime}(object, ...)

\S4method{.setPars}{Regime,numeric}(object, pars)

\S4method{show}{Regime}(object)

\S4method{summary}{Regime}(object, ...)
}
\description{
Methods Available for Objects of Class \code{Regime}

\code{.getNumPars}
  retrieves the number of parameters in the regime to be estimated.

\code{.getParNames}
  retrieves the parameter names in the regime.

\code{.getPars}
  retrieves current estimates for regime parameters.

\code{.getRegimeFunction}
  retrieves the user specified function definition of the regime.

\code{.predictOptimalTx}
  executes user specified function using current estimated parameters and
  provided data to determine recommended tx.

\code{print}
  prints the current estimates for the regime parameters.

\code{regimeCoef}
  retrieves the current estimates for the regime parameters

\code{.setPars}
  sets the parameter estimates to the provided values.

\code{show}
  displays the current estimates for the regime parameters.

\code{summary}
  retrieves the current estimates for the regime parameters
}
\keyword{internal}
