% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_Regime.R
\name{regimeCoef}
\alias{regimeCoef}
\title{Extract Regime Parameters}
\usage{
regimeCoef(object, ...)
}
\arguments{
\item{object}{A value object returned by a statistical method of DynTxRegime.}

\item{...}{Ignored.}
}
\description{
Extract the estimated regime parameters.
}
\details{
Methods are defined for all statistical methods implemented in DynTxRegime
  that use a non-regression based regime. Specifically, OptimalSeq, OWL,
  BOWL, RWL, and EARL.
}
