% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_RegimeObj.R
\docType{class}
\name{RegimeObj-class}
\alias{RegimeObj-class}
\title{Class \code{RegimeObj}}
\description{
Class \code{RegimeObj} holds information regarding regimes communicated 
  through functions under a common name.
}
\section{Slots}{

\describe{
\item{\code{regime}}{ANY expected to be \code{Regime} or \code{DecisionPointList}}
}}

\keyword{internal}
