% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_RegimeObj.R
\name{RegimeObj-methods}
\alias{RegimeObj-methods}
\alias{.getNumPars,RegimeObj-method}
\alias{.getParNames,RegimeObj-method}
\alias{.getPars,RegimeObj-method}
\alias{.getRegimeFunction,RegimeObj-method}
\alias{.predictOptimalTx,RegimeObj,data.frame-method}
\alias{print,RegimeObj-method}
\alias{regimeCoef,RegimeObj-method}
\alias{.setPars,RegimeObj,numeric-method}
\alias{show,RegimeObj-method}
\alias{summary,RegimeObj-method}
\title{Methods Available for Objects of Class \code{RegimeObj}}
\usage{
\S4method{.getNumPars}{RegimeObj}(object)

\S4method{.getParNames}{RegimeObj}(object)

\S4method{.getPars}{RegimeObj}(object)

\S4method{.getRegimeFunction}{RegimeObj}(object)

\S4method{.predictOptimalTx}{RegimeObj,data.frame}(x, newdata, dp = 1L, ...)

\S4method{print}{RegimeObj}(x, ...)

\S4method{regimeCoef}{RegimeObj}(object, ...)

\S4method{.setPars}{RegimeObj,numeric}(object, pars)

\S4method{show}{RegimeObj}(object)

\S4method{summary}{RegimeObj}(object, ...)
}
\description{
Methods dispatch equivalantly named functions defined for Regime or
  DecisionPointList objects. Method dispatched dictated by object stored
  in @regime.
}
\keyword{internal}
