\name{RegimeObject-class}
\Rdversion{1.1}
\docType{class}
\alias{RegimeObject-class}

\title{Class \code{"RegimeObject"}}

\description{
Class to identify an object as regime class.
}

\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}

\section{Methods}{
No methods defined with class \code{"RegimeObject"} in the signature.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("RegimeObject")
}

\keyword{internal}
