% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_generics.R, R/E_class_IQLearnFS_C.R,
%   R/E_class_IQLearnFS_VHet.R
\name{residuals}
\alias{residuals}
\alias{residuals,IQLearnFS_C-method}
\alias{residuals,IQLearnFS_VHet-method}
\title{Extract Model Residuals}
\usage{
residuals(object, ...)

\S4method{residuals}{IQLearnFS_C}(object, ...)

\S4method{residuals}{IQLearnFS_VHet}(object, ...)
}
\arguments{
\item{object}{A value object returned by iqLearnC() or iqLearnVar()}

\item{...}{Ignored.}
}
\description{
Retrieve residuals from an interactive Q-Learning step.
}
