% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnFS_C.R
\docType{methods}
\name{sd}
\alias{sd}
\alias{sd,IQLearnFS_C-method}
\title{Standard Deviation}
\usage{
sd(x, na.rm=FALSE)

  ## S4 method for IQLearnFS_C
  sd(x, na.rm=FALSE)
}
\arguments{
\item{x}{An object of class \code{IQLearnFS_C}}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Retrieve the standard deviation of the residuals for the first-stage contrasts
  regression in the interactive Q-Learning algorithm.
}
