% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_class_OptimalSeq.R,
%   R/H_class_OptimalSeqCoarsened.R, R/H_class_OptimalSeqMissing.R
\name{seqFunc}
\alias{seqFunc}
\alias{.seqFunc,numeric,TxInfoList-method}
\alias{.seqFunc,numeric,TxObj-method}
\title{Define the Objective Function}
\usage{
.seqFunc(eta, txObj, ...)

\S4method{.seqFunc}{numeric,TxInfoList}(eta, txObj, regimesObj, l.data, outcomeObj, propenObj, response)

\S4method{.seqFunc}{numeric,TxObj}(eta, txObj, regimesObj, l.data, outcomeObj, propenObj, response)
}
\description{
Method is defined by inheriting classes to define the objective function
  optmized by the genetic algorithm.
}
\keyword{internal}
