\name{SingleDecisionPoint-class}
\docType{class}
\alias{SingleDecisionPoint-class}

\title{Class \code{"SingleDecisionPoint"}}

\description{
A class to indicate that an object is for a single decision point.
}

\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}

\section{Methods}{
  No methods exist for this class.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("SingleDecisionPoint")
}

\keyword{internal}
