% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_SmoothRampSurrogate.R
\docType{class}
\name{SmoothRampSurrogate-class}
\alias{SmoothRampSurrogate-class}
\title{Class \code{SmoothRampSurrogate}}
\description{
Components of smoothed ramp surrogate for 0/1 loss function.
}
\keyword{internal}
