% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_SmoothRampSurrogate.R
\name{SmoothRampSurrogate-methods}
\alias{SmoothRampSurrogate-methods}
\alias{.phiFunc,SmoothRampSurrogate-method}
\alias{.dPhiFunc,SmoothRampSurrogate-method}
\title{Methods Available for Objects of Class \code{SmoothRampSurrogate}}
\usage{
\S4method{.phiFunc}{SmoothRampSurrogate}(surrogate, u, res)

\S4method{.dPhiFunc}{SmoothRampSurrogate}(surrogate, u, du, res)
}
\description{
Methods Available for Objects of Class \code{SmoothRampSurrogate}

\code{.phiFunc}
  calculates smoothed ramp surrogate loss-function

\code{.dphiFunc}
  calculates derivative of smoothed ramp surrogate loss-function
}
\keyword{internal}
