% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_SqHingeSurrogate.R
\docType{methods}
\name{SqHingeSurrogate-methods}
\alias{SqHingeSurrogate-methods}
\alias{.phiFunc,SqHingeSurrogate-method}
\alias{.dPhiFunc,SqHingeSurrogate-method}
\title{Methods Available for Objects of Class \code{SqHingeSurrogate}}
\usage{
\S4method{.phiFunc}{SqHingeSurrogate}(surrogate, u)

\S4method{.dPhiFunc}{SqHingeSurrogate}(surrogate, u, du)
}
\description{
Methods Available for Objects of Class \code{SqHingeSurrogate}

\code{.phiFunc}
  calculates squared hinge surrogate loss-function

\code{.dphiFunc}
  calculates derivative of squared hinge surrogate loss-function
}
\keyword{internal}
