\name{SubsetListFit_DecisionPointList-class}
\docType{class}
\alias{SubsetListFit_DecisionPointList-class}

\title{Class \code{"SubsetListFit_DecisionPointList"}}

\description{
A list of \code{"\link{SubsetListFit-class}"} objects, one for each decision point.}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("SubsetListFit_DecisionPointList", ...)}.
These objects are for convenience in package development
and should not be created by users.

}

\section{Extends}{
Classes \code{"\link{DecisionPointList-class}"}, 
\code{"\link{OutcomeRegression-class}"}, and
\code{"\link{SubsetsModeled-class}"}, directly.
Class \code{"\link{List-class}"}, by class \code{"\link{DecisionPointList-class}"}, distance 2.
}

\section{Methods}{
  No method identified with signature \code{"SubsetListFit_DecisionPointList-class"}.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("SubsetListFit_DecisionPointList")
}

\keyword{internal}
