\name{SubsetsModeled-class}
\Rdversion{1.1}
\docType{class}
\alias{SubsetsModeled-class}

\title{Class \code{"SubsetsModeled"}}

\description{
Class to identify a regression used subset modeling.
}

\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}

\section{Methods}{
No methods defined with class \code{"SubsetsModeled"} in the signature.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("SubsetsModeled")
}

\keyword{internal}
