\name{SubsetsNotModeled-class}
\Rdversion{1.1}
\docType{class}
\alias{SubsetsNotModeled-class}

\title{Class \code{"SubsetsNotModeled"}}

\description{
Class to identify that a regression do not use subset modeling.
}

\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}

\section{Methods}{
No methods defined with class \code{"SubsetsNotModeled"} in the signature.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("SubsetsNotModeled")
}

\keyword{internal}
