% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_generics.R
\name{summary}
\alias{summary}
\title{Result Summaries}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{Value object returned by a statistical method}

\item{...}{Optional additional inputs}
}
\description{
Returns a list of the primary results, including regression results, 
  optimization results, estimated tx and value, etc.
}
\details{
Methods are defined for all statistical methods implemented in DynTxRegime.

The exact structure of the returned list will vary depending on
 the statistical method.
}
