% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_Surrogate.R
\docType{class}
\name{Surrogate-class}
\alias{Surrogate-class}
\title{Class \code{Surrogate}}
\description{
General class for surrogate objects.
}
\section{Slots}{

\describe{
\item{\code{we}}{included to avoid VIRTUAL designation}
}}

\keyword{internal}
