% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_Surrogate.R
\name{Surrogate-methods}
\alias{Surrogate-methods}
\alias{.optim,Surrogate-method}
\title{Methods Available for Objects of Class \code{Surrogate}}
\usage{
\S4method{.optim}{Surrogate}(surrogate, par, lambda, fn, gr, suppress, ...)
}
\description{
Utilizes stats::optim to obtain parameter estimates. Requires that
the objective function and its derivative are defined by the
calling learning method. Returns NULL if optimization is not successful 
due to problems;
a vector of the current parameter estimates if optimization is not
successful because it hit the maximum number if iterations; and
the list object returned by stats::optim if optimization is successful
}
\keyword{internal}
