% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxInfoBasic.R
\docType{class}
\name{TxInfoBasic-class}
\alias{TxInfoBasic-class}
\title{Class \code{TxInfoBasic}}
\description{
Class \code{TxInfoBasic} stores basic treatment information.
}
\section{Slots}{

\describe{
\item{\code{superset}}{A vector of all possible tx options.}

\item{\code{txName}}{A character - column header of data.frame that contains tx 
variable}
}}

\keyword{internal}
