% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxInfoFactor.R
\name{TxInfoFactor-methods}
\alias{TxInfoFactor-methods}
\alias{.compareTx,TxInfoFactor,factor,factor-method}
\alias{.convertFromBinary,TxInfoFactor-method}
\alias{.convertTx,TxInfoFactor-method}
\alias{.getLevels,TxInfoFactor,factor-method}
\title{Methods Available for Objects of Class \code{TxInfoFactor}}
\usage{
\S4method{.compareTx}{TxInfoFactor,factor,factor}(object, vec1, vec2)

\S4method{.convertFromBinary}{TxInfoFactor}(txObj, txVec, ...)

\S4method{.convertTx}{TxInfoFactor}(object, txVec)

\S4method{.getLevels}{TxInfoFactor,factor}(object, txVec)
}
\description{
Methods Available for Objects of Class \code{TxInfoFactor}

\code{.compareTx(object, vec1, vec2)}
  compares vec1 and vec2 to identify equivalent elements.

\code{.convertFromBinary(txObj, txVec)}
  converts a -1/1 Tx to user provided tx coding

\code{.compareTx(object, vec1, vec2)}
  converts txVec to factor.

\code{.getLevels(object, txVec)}
  determines tx levels contains in txVec.
}
\keyword{internal}
