% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxInfoInteger.R
\name{TxInfoInteger-methods}
\alias{TxInfoInteger-methods}
\alias{.compareTx,TxInfoInteger,integer,integer-method}
\alias{.convertFromBinary,TxInfoInteger-method}
\alias{.convertTx,TxInfoInteger-method}
\alias{.getLevels,TxInfoInteger,integer-method}
\title{Methods Available for Objects of Class \code{TxInfoInteger}}
\usage{
\S4method{.compareTx}{TxInfoInteger,integer,integer}(object, vec1, vec2)

\S4method{.convertFromBinary}{TxInfoInteger}(txObj, txVec, ...)

\S4method{.convertTx}{TxInfoInteger}(object, txVec)

\S4method{.getLevels}{TxInfoInteger,integer}(object, txVec)
}
\description{
Methods Available for Objects of Class \code{TxInfoInteger}

\code{.compareTx(object, vec1, vec2)}
  compares vec1 and vec2 to identify equivalent elements.

\code{.convertFromBinary(txObj, txVec)}
  converts a -1/1 Tx to user provided tx coding

\code{.compareTx(object, vec1, vec2)}
  converts txVec to factor.

\code{.getLevels(object, txVec)}
  determines tx levels contains in txVec.
}
\keyword{internal}
