% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxInfoList.R
\docType{methods}
\name{TxInfoList-methods}
\alias{TxInfoList-methods}
\alias{.getPtsSubset,TxInfoList-method}
\alias{.getSingleton,TxInfoList-method}
\alias{.getSubsetRule,TxInfoList-method}
\alias{.getSubsets,TxInfoList-method}
\alias{.getSuperset,TxInfoList-method}
\alias{.getTxName,TxInfoList-method}
\alias{.validTx,TxInfoList-method}
\alias{.compareTx,TxInfoList,ANY,ANY-method}
\alias{.convertTx,TxInfoList-method}
\alias{.getLevels,TxInfoList,ANY-method}
\title{Methods Available for Objects of Class \code{TxInfoList}}
\usage{
\S4method{.getPtsSubset}{TxInfoList}(object)

\S4method{.getSingleton}{TxInfoList}(object)

\S4method{.getSubsetRule}{TxInfoList}(object)

\S4method{.getSubsets}{TxInfoList}(object)

\S4method{.getSuperset}{TxInfoList}(object)

\S4method{.getTxName}{TxInfoList}(object)

\S4method{.validTx}{TxInfoList}(object)

\S4method{.compareTx}{TxInfoList,ANY,ANY}(object)

\S4method{.convertTx}{TxInfoList}(object)

\S4method{.getLevels}{TxInfoList,ANY}(object)
}
\description{
Methods Available for Objects of Class \code{TxInfoList}

\code{.getPtsSubset(object)}
  not allowed.

\code{.getSingleton(object)}
  not allowed.

\code{.getSubsetRule(object)}
  not allowed.

\code{.getSubsets(object)}
  not allowed.

\code{.getSuperset(object)}
  not allowed.

\code{.getTxName(object)}
  not allowed.

\code{.validTx(object, txVec)}
  not allowed.

\code{.compareTx(object, vec1, vec2)}
  not allowed.

\code{.convertTx(object, txVec)}
  not allowed.

\code{.getLevels(object, txVec)}
  not allowed.
}
\keyword{internal}
