% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxInfoList.R
\docType{class}
\name{TxInfoList}
\alias{TxInfoList}
\title{Class \code{TxInfoList}}
\description{
Class \code{TxInfoList} extends class \code{TxObj} to indicate that
  \@txInfo is of class \code{List}. Each element of that list corresponds
  to a decision point. All methods called with this object throw errors.
}
\section{Slots}{

\describe{
\item{\code{txInfo}}{A List object}
}}

\keyword{internal}
