% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxInfoNoSubsets.R
\docType{class}
\name{TxInfoNoSubsets-class}
\alias{TxInfoNoSubsets-class}
\title{Class \code{TxInfoNoSubsets}}
\description{
Class \code{TxInfoNoSubsets} extends class \code{TxObj} to indicate that
\@txInfo is of class \code{TxInfoBasic} and thus no subsets were identified.
}
\section{Slots}{

\describe{
\item{\code{txInfo}}{A TxInfoBasic object}
}}

\keyword{internal}
