% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxInfoWithSubsets.R
\name{TxInfoWithSubsets-methods}
\alias{TxInfoWithSubsets-methods}
\alias{.getPtsSubset,TxInfoWithSubsets-method}
\alias{.getSingleton,TxInfoWithSubsets-method}
\alias{.getSubsetRule,TxInfoWithSubsets-method}
\alias{.getSubsets,TxInfoWithSubsets-method}
\title{Methods Available for Objects of Class \code{TxInfoWithSubsets}}
\usage{
\S4method{.getPtsSubset}{TxInfoWithSubsets}(object)

\S4method{.getSingleton}{TxInfoWithSubsets}(object)

\S4method{.getSubsetRule}{TxInfoWithSubsets}(object)

\S4method{.getSubsets}{TxInfoWithSubsets}(object)
}
\description{
Methods Available for Objects of Class \code{TxInfoWithSubsets}

\code{.getPtsSubset(object)}
  retrieves subset name to which each pt is a member. Method dispatched
  depends on class of @txInfo.

\code{.getSingleton(object)}
  retrieves T/F indicating if >1 tx is available to each pt. Method dispatched
  depends on class of @txInfo.

\code{.getSubsetRule(object)}
  retrieves feasible tx function. Method dispatched
  depends on class of @txInfo.

\code{.getSubsets(object)}
  retrieves feasible tx information. Method dispatched
  depends on class of @txInfo.
}
\keyword{internal}
