% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxObj.R
\docType{class}
\name{TxObj-class}
\alias{TxObj-class}
\title{Class \code{TxObj}}
\description{
Storage Class to group tx information under a common name.
}
\section{Slots}{

\describe{
\item{\code{txInfo}}{Any object -- expected to be of class TxInfoBasic,
TxInfoSubset, or DecisionPointList}
}}

\keyword{internal}
