% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxObj.R
\docType{methods}
\name{TxObj-methods}
\alias{TxObj-methods}
\alias{.compareTx,TxObj,ANY,ANY-method}
\alias{.convertFromBinary,TxObj-method}
\alias{.convertToBinary,TxObj-method}
\alias{.convertTx,TxObj-method}
\alias{.getLevels,TxObj,ANY-method}
\alias{.getSuperset,TxObj-method}
\alias{.getTxName,TxObj-method}
\alias{.validTx,TxObj-method}
\title{Methods Available for Objects of Class \code{TxObj}}
\usage{
\S4method{.compareTx}{TxObj,ANY,ANY}(object, vec1, vec2)

\S4method{.convertFromBinary}{TxObj}(txObj, txVec, ...)

\S4method{.convertToBinary}{TxObj}(txObj, data, ...)

\S4method{.convertTx}{TxObj}(object, txVec)

\S4method{.getLevels}{TxObj,ANY}(object, txVec)

\S4method{.getSuperset}{TxObj}(object)

\S4method{.getTxName}{TxObj}(object)

\S4method{.validTx}{TxObj}(object, txVec)
}
\description{
Methods dispatched depend on class of @txInfo.
}
\keyword{internal}
