% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxSubsetFactor.R
\docType{class}
\name{TxSubsetFactor-class}
\alias{TxSubsetFactor-class}
\title{Class \code{TxSubsetFactor}}
\description{
Class \code{TxSubsetFactor} stores subset information for tx when tx is
  a factor
}
\keyword{internal}
