% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxSubsetInteger.R
\docType{class}
\name{TxSubsetInteger-class}
\alias{TxSubsetInteger-class}
\title{Class TxSubsetInteger}
\description{
Class \code{TxSubsetInteger} stores subset information for treatment
}
\keyword{internal}
