% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_TypedFit.R
\docType{class}
\name{TypedFit-class}
\alias{TypedFit-class}
\title{Class \code{TypedFit}}
\description{
Class \code{TypedFit} is a \code{modelObjFit} combined with a character
  to identify its purpose.
}
\keyword{internal}
