% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_TypedFit.R
\docType{methods}
\name{TypedFit-methods}
\alias{TypedFit-methods}
\alias{coef,TypedFit-method}
\alias{fitObject,TypedFit-method}
\alias{plot,TypedFit,ANY-method}
\alias{print,TypedFit-method}
\alias{show,TypedFit-method}
\alias{summary,TypedFit-method}
\title{Methods Available for Objects of Class \code{TypedFit}}
\usage{
\S4method{coef}{TypedFit}(object, ...)

\S4method{fitObject}{TypedFit}(object, ...)

\S4method{plot}{TypedFit,ANY}(x, suppress = FALSE, ...)

\S4method{print}{TypedFit}(x, ...)

\S4method{show}{TypedFit}(object)

\S4method{summary}{TypedFit}(object, ...)
}
\description{
Methods call equivalently named methods defined for \code{modelObjFit}
  objects.
}
\keyword{internal}
