% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_TypedFit_fSet.R
\docType{class}
\name{TypedFit_fSet-class}
\alias{TypedFit_fSet-class}
\title{Class \code{TypedFit_fSet}}
\description{
Class \code{TypedFit_fSet} is a \code{TypedFit} when subsets are identified
  but not modeled independently.
}
\keyword{internal}
