% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_TypedFit_fSet.R
\name{TypedFit_fSet-methods}
\alias{TypedFit_fSet-methods}
\alias{coef,TypedFit_fSet-method}
\alias{fitObject,TypedFit_fSet-method}
\alias{plot,TypedFit_fSet,ANY-method}
\alias{predict,TypedFit_fSet-method}
\alias{print,TypedFit_fSet-method}
\alias{show,TypedFit_fSet-method}
\alias{summary,TypedFit_fSet-method}
\title{Methods Available for Objects of Class \code{TypedFit_fSet}}
\usage{
\S4method{coef}{TypedFit_fSet}(object, ...)

\S4method{fitObject}{TypedFit_fSet}(object, ...)

\S4method{plot}{TypedFit_fSet,ANY}(x, suppress = FALSE, ...)

\S4method{predict}{TypedFit_fSet}(object, newdata, ...)

\S4method{print}{TypedFit_fSet}(x, ...)

\S4method{show}{TypedFit_fSet}(object)

\S4method{summary}{TypedFit_fSet}(object, ...)
}
\description{
Methods call equivalently named methods defined for \code{TypedFit}
  objects.

\code{predict(object, ...)}
  Patients with only 1 tx option are NA.
}
\keyword{internal}
