% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_TypedFit_SubsetList.R
\docType{class}
\name{TypedFit_SubsetList-class}
\alias{TypedFit_SubsetList-class}
\title{Class \code{TypedFit_SubsetList}}
\description{
Class \code{TypedFit_SubsetList} is \code{SubsetList} of \code{TypedFit} 
used when subsets are identified and modeled independently.
}
\keyword{internal}
