% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_TypedFit_SubsetList.R
\name{TypedFit_SubsetList-methods}
\alias{TypedFit_SubsetList-methods}
\alias{coef,TypedFit_SubsetList-method}
\alias{fitObject,TypedFit_SubsetList-method}
\alias{plot,TypedFit_SubsetList,ANY-method}
\alias{predict,TypedFit_SubsetList-method}
\alias{print,TypedFit_SubsetList-method}
\alias{show,TypedFit_SubsetList-method}
\alias{summary,TypedFit_SubsetList-method}
\title{Methods Available for Objects of Class \code{TypedFit_SubsetList}}
\usage{
\S4method{coef}{TypedFit_SubsetList}(object, ...)

\S4method{fitObject}{TypedFit_SubsetList}(object, ...)

\S4method{plot}{TypedFit_SubsetList,ANY}(x, suppress = FALSE, ...)

\S4method{predict}{TypedFit_SubsetList}(object, newdata, ...)

\S4method{print}{TypedFit_SubsetList}(x, ...)

\S4method{show}{TypedFit_SubsetList}(object)

\S4method{summary}{TypedFit_SubsetList}(object, ...)
}
\description{
Methods call equivalently named methods defined for \code{TypedFit}
  objects. When a value object is returned, it is a list. The element
  names of the list are the subset names to which the result pertains.

\code{predict(object, ...)}
  Patients not in subset are NA.
}
\keyword{internal}
