% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_TypedFitObj.R
\docType{class}
\name{TypedFitObj-class}
\alias{TypedFitObj-class}
\title{Class \code{TypedFitObj}}
\description{
Class \code{TypedFit_SubsetList} allows for TypedFit based objects to be 
grouped under a common name.
}
\section{Slots}{

\describe{
\item{\code{fit}}{ANY - expected to be \code{TypedFit}, \code{TypedFit_fSet}.
\code{TypedFit_SubsetList} or \code{DecisionPointList} of these.}
}}

\keyword{internal}
