% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_TypedFitObj.R
\name{TypedFitObj-methods}
\alias{TypedFitObj-methods}
\alias{coef,TypedFitObj-method}
\alias{fitObject,TypedFitObj-method}
\alias{plot,TypedFitObj,ANY-method}
\alias{predict,TypedFitObj-method}
\alias{print,TypedFitObj-method}
\alias{show,TypedFitObj-method}
\alias{summary,TypedFitObj-method}
\title{Methods Available for Objects of Class \code{TypedFitObj}}
\usage{
\S4method{coef}{TypedFitObj}(object, ...)

\S4method{fitObject}{TypedFitObj}(object, ...)

\S4method{plot}{TypedFitObj,ANY}(x, suppress = FALSE, ...)

\S4method{predict}{TypedFitObj}(object, ...)

\S4method{print}{TypedFitObj}(x, ...)

\S4method{show}{TypedFitObj}(object)

\S4method{summary}{TypedFitObj}(object, ...)
}
\description{
Methods call equivalently named methods defined for \code{TypedFit},
  \code{TypedFit_fSet}, \code{TypedFit_SubsetList} or
  \code{DecisionPointList} objects.
The methods dispatched and objects returned depend on the class of @fit.
}
\keyword{internal}
