\name{TypedSimpleFitNoSubsets-class}
\docType{class}
\alias{TypedSimpleFitNoSubsets-class}

\title{Class \code{"TypedSimpleFitNoSubsets"}}

\description{
An outcome regression step with a single model and no
subset modeling.
}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("TypedSimpleFitNoSubsets", ...)}.
These objects are for convenience in package development
and should not be created by users.
}

\section{Extends}{
Class \code{"\link{TypedSimpleFit-class}"} directly.
Classes \code{"modelObjFit"}, \code{"\link{SingleDecisionPoint-class}"}, 
\code{"\link{OutcomeRegression-class}"}, and \code{"\link{SubsetsNotModeled-class}"}
by class \code{"\link{TypedSimpleFit-class}"} distance 2.
}

\section{Methods}{
No methods defined with class \code{"TypedSimpleFitNoSubsets"} in the signature.
}

\author{
Shannon T Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("TypedSimpleFitNoSubsets")
}

\keyword{internal}

