\name{bmiData}
\docType{data}
\alias{bmiData}
\title{
Adolescent BMI dataset (generated toy example)
}
\description{
A dataset generated to mimic data from a two-stage randomized clinical trial 
that studied the effect of meal replacement shakes on adolescent obesity. 
The dataset contains the following covariates collected at the start of the 
first stage: "gender," "race," "parentBMI," and "baselineBMI." At the 
second-stage, "month4BMI" was collected. Variables "A1" and "A2" are the 
randomized treatments at stages one and two, and "month12BMI" is the primary 
outcome collected at the end of stage two. 
}
\usage{
bmiData
}
\format{
A matrix with rows corresponding to patients.
}
\source{
Generated by Kristin A. Linn in R
}
\keyword{datasets }
