% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_class_BOWL.R
\docType{class}
\name{BOWL-class}
\alias{BOWL-class}
\title{Class \code{BOWL}}
\description{
Class \code{BOWL} contains results from a single step of BOWL algorithm.
}
\section{Slots}{

\describe{
\item{\code{step}}{Integer indicating step of the algorithm}

\item{\code{analysis}}{Contains a Learning or LearningMulti object.}

\item{\code{analysis@txInfo}}{Feasible tx information.}

\item{\code{analysis@propen}}{Propensity regression analysis.}

\item{\code{analysis@outcome}}{Outcome regression analysis.}

\item{\code{analysis@cvInfo}}{Cross-validation analysis if single regime.}

\item{\code{analysis@optim}}{Optimization analysis if single regime.}

\item{\code{analysis@optimResult}}{list of cross-validation and optimization results
if multiple regimes. optimResult[[i]]@cvInfo and optimResult[[i]]@optim.}

\item{\code{analysis@optimal}}{Estimated optimal Tx and value.}

\item{\code{analysis@call}}{Unevaluated call to statistical method.}

\item{\code{prodPi}}{Vector of the products of the propensity for the tx received}

\item{\code{sumR}}{Vector of the sum of the rewards}

\item{\code{index}}{Vector indicating compliance with estimated optimal regime}
}}

\keyword{internal}
