% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_class_BOWLBasic.R
\docType{class}
\name{BOWLObj-class}
\alias{BOWLObj-class}
\title{Class \code{BOWLObj}}
\description{
Class \code{BOWLObj} contains product and sum information required for 
  iteration
}
\section{Slots}{

\describe{
\item{\code{prodPi}}{Vector of the products of the propensity for the tx received}

\item{\code{sumR}}{Vector of the sum of the rewards}

\item{\code{index}}{Vector indicating compliance with estimated optimal regime}
}}

\keyword{internal}
