% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVInfo.R
\docType{class}
\name{CVInfo-class}
\alias{CVInfo-class}
\title{Class \code{CVInfo}}
\description{
Class \code{CVInfo} holds cross-validation procedure results
}
\section{Slots}{

\describe{
\item{\code{value}}{Values obtained for each parameter combination}

\item{\code{params}}{list of parameter values considered}

\item{\code{optimal}}{list of optimal parameter values}
}}

\keyword{internal}
