% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_CVInfo2Par.R
\docType{class}
\name{CVInfo2Par-class}
\alias{CVInfo2Par-class}
\title{Class \code{CVInfo2Par}}
\description{
Class \code{CVInfo2Par} holds information regarding cross-validation
 procedure when multiple kernel parameters and tuning parameters are 
 considered.
}
\section{Slots}{

\describe{
\item{\code{value}}{Matrix of values at parameters considered}
}}

\keyword{internal}
