% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_DecisionPointList.R
\docType{class}
\name{DecisionPointList-class}
\alias{DecisionPointList-class}
\alias{initialize,DecisionPointList-method}
\title{Class \code{DecisionPointList}}
\usage{
\S4method{initialize}{DecisionPointList}(.Object, ...)
}
\description{
Class \code{DecisionListList} represents a \code{List} for decision points.
This class extends \code{List} to require non-zero length.
}
\keyword{internal}
