% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_DynTxRegime.R
\docType{class}
\name{DynTxRegime-class}
\alias{DynTxRegime-class}
\title{Class \code{DynTxRegime}}
\description{
Class \code{DynTxRegime} is a component of all statistical methods 
implemented in the package. This class contains the estimated optimal Tx, 
decision functions if applicable, the estimated value and the original 
unevaluated call. It extends internal class \code{OptimalObj}.
}
\section{Slots}{

\describe{
\item{\code{call}}{object of class call or NULL}
}}

\keyword{internal}
