\name{internals}
\alias{[[,List-method}
\alias{[[<-,QLearnEstList-method}
\alias{[[<-,List-method}
\alias{length,List-method}
\alias{qFuncs,IQEst-method}
\alias{qFuncs,IQLearnFS_VHom-method}
\alias{qFuncs,QLearnEst-method}
\alias{qFuncs,OptimalSeq-method}
\alias{qFuncs,OptimalClass-method}
\alias{qqPlot,IQLearnFS_VHet-method}


\title{ ~~ }
\description{~~}
\keyword{internal}
