\name{IQLearnFS-class}
\Rdversion{1.1}
\docType{class}
\alias{IQLearnFS-class}

\title{Class \code{"IQLearnFS"}}
\description{
A VIRTUAL class used to specify an object as corresponding to
a first-stage IQ-Learning analysis.
}

\section{Objects from the Class}{
A virtual Class: No objects may be created from it.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\keyword{internal}

