% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_class_IQLearnFS_C.R
\docType{class}
\name{IQLearnFS_C-class}
\alias{IQLearnFS_C-class}
\title{Class \code{IQLearnFS_C}}
\description{
Class \code{IQLearnFS_C} contains the results for the first stage
 contrasts component of the interactive Q-Learning algorithm.
 Objects of this class are returned by iqLearnFSC().
}
\section{Slots}{

\describe{
\item{\code{txVec}}{: A numeric. treatment vector from training data}

\item{\code{residuals}}{: A numeric. residuals of the fit}

\item{\code{step}}{: Not used in this context.}

\item{\code{outcome}}{: The outcome regression analysis}

\item{\code{txInfo}}{: The feasible tx information}

\item{\code{optimal}}{: The estimated optimal tx, decision function, and value}
}}

\section{Methods For Post-Processing of Regression Analysis}{


 \describe{
   \item{outcome}{: 
      Retrieve value object returned by outcome regression methods.  }
}



 \describe{
   \item{coef}{: 
     Retrieve parameter estimates for all regression steps. }
   \item{fitObject}{: 
      Retrieve value object returned by regression methods. }
   \item{plot}{: 
      Generate plots for regression analyses.  }
}
}

\section{Methods For Accessing Main Results}{


 \describe{
   \item{DTRstep}{: 
      Retrieve description of method used to create object. }
   \item{estimator}{: 
      Retrieve the estimated value of the estimated
      optimal regime for the training data set. }
   \item{optTx}{: 
      Retrieve/predict the estimated decision functions and/or optimal tx. }
   \item{print}{: 
      Print main results of analysis.}
   \item{show}{: 
      Show main results of analysis.}
   \item{summary}{: 
      Retrieve summary information.}
}


\describe{
   \item{residuals}{:Retrieve the residuals of the regression.}
   \item{sd}{:Retrieve the standard deviation of the residuals.}
}
}

