\name{fittedCont}
\alias{fittedCont}
\alias{fittedCont,DynTxRegime-method}

\title{ Extract Fitted Contrast Functions}
\description{
Extract fitted contrast functions for IQ-Learning algorithm.
}

\usage{ \S4method{fittedCont}{DynTxRegime}(object, ...)  }

\arguments{
  \item{object }{an object of class DynTxRegime.}

  \item{\dots }{ignored.}

}

\value{
For object returned by IQ-Learning functions, the fitted contrast functions.
For all others, NULL
}

\author{
  Shannon T. Holloway <sthollow@ncsu.edu>
}


