\name{residuals}
\alias{residuals}
\alias{residuals,DynTxRegime-method}

\title{ Extract Model Residuals}
\description{
Extract model residuals.
}

\usage{ \S4method{residuals}{DynTxRegime}(object, ...)  }

\arguments{
  \item{object }{an object of class DynTxRegime.}

  \item{\dots }{ignored.}

}

\value{
Residuals extracted from the DynTxRegime object. For outcome
regression, these will be
the residuals for the full/combined model.
}

\author{
  Shannon T. Holloway <sthollow@ncsu.edu>
}


