\name{stdDev}
\alias{stdDev}
\alias{stdDev,DynTxRegime-method}

\title{Standard Deviation of IQ-Learning Variance Step}

\description{Returns the standard deviation from the IQ-Learning
homoskedastic variance calculation.}

\usage{ \S4method{stdDev}{DynTxRegime}(object, ...)  }

\arguments{
  \item{object }{an object of class DynTxRegime.}

  \item{\dots }{ignored.}

}

\value{
A numeric object.
}

\author{
  Shannon T. Holloway <sthollow@ncsu.edu>
}

