\name{summary}
\docType{methods}
\alias{summary-methods}
\alias{summary}
\alias{summary,DynTxRegime-method}

\title{ Summary Results}
\description{
Generate summary results for a DynTxRegime method.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "DynTxRegime")}}{
Generate summary results for all regression steps.
}

}}
\keyword{methods}
