% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_ModelObj_DecisionPointList.R
\docType{class}
\name{ModelObj_DecisionPointList-class}
\alias{ModelObj_DecisionPointList-class}
\alias{ModelObj_DecisionPointList}
\title{Class ModelObj_DecisionPointList}
\usage{
\S4method{initialize}{ModelObj_DecisionPointList}(.Object, ...)
}
\description{
Class \code{ModelObj_DecisionPointList} represents a \code{List} for 
multiple decision points. Contents can be other modelObj or ModeObj_SubsetList.
}
\keyword{internal}
