\name{ModelObj_SubsetList-class}
\docType{class}
\alias{ModelObj_SubsetList-class}

\title{Class \code{"ModelObj_SubsetList"}}

\description{
A list of objects of class \code{"modelObj"} defined for subsets.
}

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ModelObj_SubsetList", ...)}.
These objects are for convenience in package development
and should not be created by users.
}

\section{Extends}{
Class \code{"\link{SubsetList-class}"}, directly.
Classwa \code{"\link{List-class}"} and \code{"\link{SingleDecisionPoint-class}"}, by class "SubsetList", distance 2.
}

\section{Methods}{
No methods defined with class \code{"ModelObj_SubsetList"} in the signature.
}

\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}

\examples{
showClass("ModelObj_SubsetList")
}

\keyword{internal}
