% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_ModelObjSubset.R
\docType{class}
\name{ModelObjSubset-class}
\alias{ModelObjSubset-class}
\title{Class \code{ModelObjSubset}}
\description{
Class \code{ModelObjSubset} stores a modelObj object with pertinent subset
information
}
\section{Slots}{

\describe{
\item{\code{decisionPoint}}{integer indicating the decision point for modelObj}

\item{\code{subset}}{character indicating the subset for modelObj

import modelObj}
}}

\keyword{internal}
