\name{OptimalClass-class}
\Rdversion{1.1}
\docType{class}
\alias{OptimalClass-class}

\title{Class \code{"OptimalClass"}}
\description{
VIRTUAL class for classification method with IPWE or AIPWE.
}
\section{Objects from the Class}{
A virtual Class: No object may be created from it.
}
\section{Extends}{
Class \code{"\link{DynTxRegime-class}"} directly.
}
\author{
Shannon T. Holloway <sthollow@ncsu.edu>
}
\examples{
showClass("OptimalClass")
}

\keyword{internal}

