% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_class_OptimalSeq.R
\docType{class}
\name{OptimalSeq-class}
\alias{OptimalSeq-class}
\title{Class \code{OptimalSeq}}
\description{
Class \code{OptimalSeq} contains the results for the estimated optimal tx
  and value when estimated from a coarsened or missing data perspective.
}
\section{Slots}{

\describe{
\item{\code{genetic}}{A list containing the results from the genetic algorithm}

\item{\code{propen}}{Results of the propensity regression step}

\item{\code{outcome}}{Results of the outcome regression step}

\item{\code{regime}}{Results for the regime.}

\item{\code{optimal}}{Results for the estimated optimal tx and value}

\item{\code{Call}}{The unevaluated call.}
}}

\section{Methods For Post-Processing of Regression Analysis}{


 \describe{
   \item{outcome}{: 
      Retrieve value object returned by outcome regression methods.  }
}



 \describe{
   \item{propen}{: 
      Retrieve value object returned by propensity regression methods.  }
}



 \describe{
   \item{coef}{: 
     Retrieve parameter estimates for all regression steps. }
   \item{fitObject}{: 
      Retrieve value object returned by regression methods. }
   \item{plot}{: 
      Generate plots for regression analyses.  }
}
}

\section{Methods For Accessing Main Results}{


 \describe{
   \item{regimeCoef}{: 
      Retrieve the estimated regime parameters. }
}



 \describe{
   \item{DTRstep}{: 
      Retrieve description of method used to create object. }
   \item{estimator}{: 
      Retrieve the estimated value of the estimated
      optimal regime for the training data set. }
   \item{optTx}{: 
      Retrieve/predict the estimated decision functions and/or optimal tx. }
   \item{print}{: 
      Print main results of analysis.}
   \item{show}{: 
      Show main results of analysis.}
   \item{summary}{: 
      Retrieve summary information.}
}
}

