% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_OptimBasic.R
\docType{class}
\name{OptimBasic-class}
\alias{OptimBasic-class}
\title{Class \code{OptimBasic}}
\description{
Class \code{OptimBasic} holds results of an optimization step when linear
  kernel is used for decision function.
}
\section{Slots}{

\describe{
\item{\code{lambda}}{A numeric, tuning parameter}

\item{\code{optim}}{A list, value object returned by optimization method
expected optimization methods are optim and hjk}

\item{\code{surrogate}}{A Surrogate object specifying loss-function surrogate}
}}

\keyword{internal}
