% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_OptimKernel.R
\docType{class}
\name{OptimKernel-class}
\alias{OptimKernel-class}
\title{Class \code{OptimKernel}}
\description{
Class \code{OptimKernel} holds results of an optimization step when non-linear
  kernel is used for decision function.
}
\keyword{internal}
