% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeNoFit.R
\docType{methods}
\name{OutcomeNoFit-methods}
\alias{OutcomeNoFit-methods}
\alias{.predictAll,OutcomeNoFit,data.frame-method}
\alias{outcome,OutcomeNoFit-method}
\alias{coef,OutcomeNoFit-method}
\alias{fitObject,OutcomeNoFit-method}
\alias{plot,OutcomeNoFit,ANY-method}
\alias{predict,OutcomeNoFit-method}
\alias{print,OutcomeNoFit-method}
\alias{show,OutcomeNoFit-method}
\alias{summary,OutcomeNoFit-method}
\title{Methods Available for Objects of Class \code{OutcomeNoFit}}
\usage{
\S4method{.predictAll}{OutcomeNoFit,data.frame}(object, newdata)

\S4method{outcome}{OutcomeNoFit}(object, ...)

\S4method{coef}{OutcomeNoFit}(object, ...)

\S4method{fitObject}{OutcomeNoFit}(object, ...)

\S4method{plot}{OutcomeNoFit,ANY}(x, suppress = FALSE, ...)

\S4method{predict}{OutcomeNoFit}(object, ...)

\S4method{print}{OutcomeNoFit}(x, ...)

\S4method{show}{OutcomeNoFit}(object)

\S4method{summary}{OutcomeNoFit}(object, ...)
}
\description{
Methods return NULL, NA or zero values.

\code{.predictAll(object, newdata)} returns a list containing the optimal
  tx as a vector of NA values and the decision function as a matrix of 0
}
\keyword{internal}
