% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_OutcomeSimpleFit.R
\docType{methods}
\name{OutcomeSimpleFit-methods}
\alias{OutcomeSimpleFit-methods}
\alias{coef,OutcomeSimpleFit-method}
\alias{fitObject,OutcomeSimpleFit-method}
\alias{outcome,OutcomeSimpleFit-method}
\alias{plot,OutcomeSimpleFit,ANY-method}
\alias{predict,OutcomeSimpleFit-method}
\alias{.predictAll,OutcomeSimpleFit,data.frame-method}
\alias{print,OutcomeSimpleFit-method}
\alias{show,OutcomeSimpleFit-method}
\alias{summary,OutcomeSimpleFit-method}
\title{Methods Available for Objects of Class \code{OutcomeSimpleFit}}
\usage{
\S4method{coef}{OutcomeSimpleFit}(object, ...)

\S4method{fitObject}{OutcomeSimpleFit}(object, ...)

\S4method{outcome}{OutcomeSimpleFit}(object, ...)

\S4method{plot}{OutcomeSimpleFit,ANY}(x, suppress = FALSE, ...)

\S4method{predict}{OutcomeSimpleFit}(object, ...)

\S4method{.predictAll}{OutcomeSimpleFit,data.frame}(object, newdata)

\S4method{print}{OutcomeSimpleFit}(x, ...)

\S4method{show}{OutcomeSimpleFit}(object)

\S4method{summary}{OutcomeSimpleFit}(object, ...)
}
\description{
Methods call equivalently named methods defined for \code{TypedFit}

\code{.predictAll(object, newdata)}
  predicts outcome for all tx options.
  Returns a list containing 'optimalTx' the tx yielding the largest 
  predicted outcome and 'decisionFunc' the matrix of outcomes predicted
  for all tx.
}
\keyword{internal}
