% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_PropensityFit.R
\docType{class}
\name{PropensityFit-class}
\alias{PropensityFit-class}
\title{Class \code{PropensityFit}}
\description{
Class \code{PropensityFit} is a \code{TypedFit} identified as being
  for a propensity regression step.
}
\section{Slots}{

\describe{
\item{\code{small}}{A logical TRUE indicates that the smallest valued tx is
missing; FALSE indicates that the largest valued tx is missing}

\item{\code{levs}}{A vector; the set of treatment options included in fit.}
}}

\keyword{internal}
