% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_Regime.R
\docType{class}
\name{Regime-class}
\alias{Regime-class}
\title{Class \code{Regime}}
\description{
Class \code{Regime} holds information regarding regimes communicated 
  through functions.
}
\section{Slots}{

\describe{
\item{\code{nVars}}{An integer. The number of parameters to be estimated}

\item{\code{vNames}}{A character. The names of the parameters to be estimated}

\item{\code{func}}{A function. The user specified function that defines the regime}

\item{\code{pars}}{A numeric. The estimated parameters}
}}

\keyword{internal}
