% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_SubsetList.R
\docType{class}
\name{SubsetList-class}
\alias{SubsetList-class}
\title{Class \code{SubsetList}}
\description{
Class \code{SubsetList} represents a \code{List} for subset specifications.
This class extends \code{List} to require non-zero length and named elements.
}
\keyword{internal}
