% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxInfoFactor.R
\docType{class}
\name{TxInfoFactor-class}
\alias{TxInfoFactor-class}
\title{Class TxInfoFactor}
\description{
Class \code{TxInfoFactor} extends \code{TxInfoBasic} to identify treatments 
as factor
}
\section{Slots}{

\describe{
\item{\code{superset}}{character of all allowed tx options}
}}

\keyword{internal}
