% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxInfoInteger.R
\docType{class}
\name{TxInfoInteger-class}
\alias{TxInfoInteger-class}
\title{Class TxInfoInteger}
\description{
Class \code{TxInfoInteger} extends \code{TxInfoBasic} to identify treatments 
as integer.
}
\keyword{internal}
