% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxInfoWithSubsets.R
\docType{class}
\name{TxInfoWithSubsets-class}
\alias{TxInfoWithSubsets-class}
\title{Class \code{TxInfoWithSubsets}}
\description{
Class \code{TxInfoWithSubsets} extends class \code{TxObj} to indicate that
\@txInfo is of class \code{TxInfoSubset} and thus subsets were identified.
}
\section{Slots}{

\describe{
\item{\code{txInfo}}{A TxSubset object}
}}

\keyword{internal}
