% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_TxSubset.R
\docType{class}
\name{TxSubset-class}
\alias{TxSubset-class}
\title{Class \code{TxSubset}}
\description{
Class \code{TxSubset} stores subset information for tx
}
\section{Slots}{

\describe{
\item{\code{ptsSubset}}{A character vector. The name of the subset of which each 
patient is a member}

\item{\code{subsetRule}}{A function. The fSet function provided by user.}

\item{\code{subsets}}{A list. The feasible treatments for each subset. The elements
must be named and contain tx subsets}

\item{\code{singleton}}{A logical vector. TRUE indicates if 1 tx is available to
each patient}
}}

\keyword{internal}
